% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{abdat}
\alias{abdat}
\title{abdat fishery data for Block 12 of Tasmanian blacklip fishery}
\format{
A 58 row x 4 column matrix
\describe{
\item{year}{ 9 fishing years from 1963 to 2020}
\item{lml}{ the legal minimum length in the year of catch}
\item{catch}{ the annual catch of blacklip abalone \emph{Haliotis rubra}}
\item{cpue}{ the standardizaed cpue for each year}
}
}
\usage{
abdat
}
\description{
A dataset containing columns of year, LML, catches, and cpue
data for use in fmr when exploring size-based integrated models.
}
\examples{
data(abdat)
abdat
}
\references{
McAllister, J. and C. Mundy (2024) \emph{Tasmanian Abalone
Assessment}, Institute for Marine & Antarctic Studies, University of
Tamsania p 1-144.
}
\keyword{datasets}
