% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{abdynamicsH}
\alias{abdynamicsH}
\title{abdynamicsH calculates the size-based dynamics for the population}
\usage{
abdynamicsH(abstk, pin, fish, const, glb, full = TRUE, initH = 0)
}
\arguments{
\item{abstk}{the stock derived from makeabstock}

\item{pin}{the log transformed vector of parameters, pindat column 1}

\item{fish}{the fishery data in two matrices. The first contains the
year, the LML, catches, and the cpue, and the second contains the
numbers-at-size in the catch, just for those years with observations.}

\item{const}{contains biological constants, including those for growth}

\item{glb}{the global object}

\item{full}{should only the predicted cpue and numbers-at-size in the
catch be output (full =FALSE), or all of the dynamics (full = TRUE)}

\item{initH}{the constant initial harvest rate that gives rise, approximately,
to the desired initial depletion. default=0, which implies no initial
depletion. default = 0}
}
\value{
only the predicted values or the full dynamics depending on full
}
\description{
abdynamicsH is the main function for calculating the size-based
population dynamics, which follows the numbers-at-size. It exports
the full dynamics or a pared down version of only the predicted cpue
and predicted numbers-at-size in the catch for the calculation of the'
minimum negative log-likelihood. This is the combination of the
Log-Normal errors on cpue and the Multinomial errors on the size-
composition of the catch. In this case it uses Rcpp and RcppArmadillo
to speed some of the computations
}
\examples{
print("wait on example data sets")
# abstk=abstk;pin=pindat[,1];fish=fish;const=const;glb=glb;full=TRUE;initH=0 
}
