% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{absaudata}
\alias{absaudata}
\title{absaudata a list of five objects relating to a Tasmanian abalone stock}
\format{
A list of five objects, both matrices and lists
\describe{
\item{fish}{58years x 4 (year, LML, catch, and cpue) matrix 1963 - 2020}
\item{sizecomp}{19 size-class x 31 year matrix of size-composition data
where some years have no data}
\item{glb}{a globals object containing model constants such as the midpts
of the size classes, number of size classes, natural mortality,
steepness, sigR, lambda (hyper-stability coefficient), and the years
in which recruitment deviates are estiamted. see str(glb)}
\item{const}{9 constants relating to weight-at, maturity-at, and
growth-at-size}
\item{pindat}{initial parameter values prior to fitting, 34 x 2 matrix,
with the second column determining which parameters are fitted}
}
}
\usage{
absaudata
}
\description{
absaudata is a list of five objects relating to a single
statistical block in teh West Coast Tasmanian blacklip abalone fishery
(Haliotis rubra). It contains fish, sizecomp, glb, const, and pindat.
These can be used to generate a stock of abalone to illustrate the use
of a size-based model (SBM).
}
\examples{
data(absaudata)
str(absaudata)
}
\keyword{datasets}
