% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspm.r
\name{aspmdynamicsF}
\alias{aspmdynamicsF}
\title{aspmdynamicsF describe the ASPM dynamics}
\usage{
aspmdynamicsF(
  pars,
  infish,
  inglb,
  inprops,
  maxF = 1,
  waa = "waa",
  maa = "maa",
  sela = "sela"
)
}
\arguments{
\item{pars}{the dynamics relies on many parameters sitting in the global
environment in particular ages, nages, maxage, M, fish, nyrs, and
maa, waa, sela, which are contained in props. 'pars' can contain either
two or three parameters. 1) is the log-transformed average unfished
recruitment, inR0. 2) is the variability around the index of relative
abundance (cpue or survey index) during the fitting process, and if is
present 3) is the initial depletion level initdepl, which if present
will be fitted as well.}

\item{infish}{the fish data.frame from readdata or built in dataset}

\item{inglb}{the glb data.frame from readdata or built in dataset}

\item{inprops}{the props data.frame from readdata or built in dataset}

\item{maxF}{the maximum instantaneous F to be used in the search for each
instantaneous F rate}

\item{waa}{the character name of the weight-at-age}

\item{maa}{the character name of the maturity-at-age}

\item{sela}{the character name of the selectivity-at-age}
}
\value{
a data.frame containing the fishery dynamics according to the input
parameter inR0. Includes Year, Catch, PredC, SpawnB, ExploitB, FullH,
CPUE, PredCE, Deplete, Recruit, FullF.
}
\description{
aspmdynamicsF summarizes the dynamics of an Age-Structured
Production Model (ASPM). Fitting the ASPM entails estimating the unfished
recruitment level (R0), which is input as a parameter. In this case
fishing mortality is implemented as instantaneous rates rather that
as annual harvest rates.
}
\examples{
\dontrun{
data(westroughy)
fish <- westroughy$fish
glb <- westroughy$glb
props <- westroughy$props
pars <- c(14,0.3)
dynamics(pars,fish,glb,props)
aspmFLL(pars,fish,glb,props)      # 
pars <- c(14.0,0.3,0.95) # logR0, sigCE, depletion
fishery <- aspmdynamicsF(pars,fish,glb,props) 
aspmFLL(pars,fish,glb,props)      # 
pars <- c(14,0.3)
bestL <- optim(pars,aspmLL,method="Nelder-Mead",infish=fish,inglb=glb,
               inprops=props,control=list(maxit = 1000,parscale = c(10,0.1)))
str(bestL)
fishery <- aspmdynamicsF(bestL$par,fish,glb,props)
print(round(fishery,4)) 
}
}
\seealso{
{
\link{dynamics}
}
}
