% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspm.r
\name{aspmindexfit}
\alias{aspmindexfit}
\title{aspmindexfit plots an aspm fit to CPUE or Index data}
\usage{
aspmindexfit(
  infish,
  glb,
  CI = NULL,
  rundir = "",
  CIlwd = 1,
  CIcol = 4,
  console = TRUE
)
}
\arguments{
\item{infish}{the matrix of fishery dynamics from the dynamicsH, dynamicsF,
of dynF functions}

\item{glb}{the globals object for the aspm analysis used}

\item{CI}{the output of the getLNCI function that defines Log-Normal CI}

\item{rundir}{default = "", give a full path if saving a png file}

\item{CIlwd}{width of CI lines default = 1}

\item{CIcol}{colour of CI lines, default = 4 = blue}

\item{console}{should the plot be sent to the console, default = TRUE. If
set = FALSE, then rundir can be set to identify where the png of the
plot will be saved}
}
\value{
the filen used, which defaults to '' but generates a plot
}
\description{
aspmindexfit plots an aspm fit to CPUE or Index data with the
option of including 95\% Log-Normal CI around the model fit.
}
\examples{
data(francis92)
glb <- francis92$glb;  props=francis92$props; fish=francis92$fish
pars <- c(11.5,-1,-1)  # Try values 11.5, 11, and 10 for the log(R0) value
out115 <- dynamicsH(pars=pars,infish=fish,inglb=glb,inprops=props,
                    full=TRUE)
pars <- c(11.0,-1,-1)
out11 <- dynamicsH(pars=pars,infish=fish,inglb=glb,inprops=props,
                   full=TRUE)
pars <- c(10.0,-1,-1)
out10 <- dynamicsH(pars=pars,infish=fish,inglb=glb,inprops=props,
                   full=TRUE)
filen <- aspmindexfit(out115$fishery,glb=glb,CI=NULL,console=TRUE)
lines(1977:1989,out11$fishery[,"predCE"],lwd=2,col=2)
lines(1977:1989,out10$fishery[,"predCE"],lwd=2,col=4)
legend("topright",c("11.5","11","10"),col=c(1,2,4),lwd=3,cex=1,bty="n")
}
\seealso{
{
\link{getLNCI}, \link{dynamicsH}, \link{dynamicsF}, \link{dynF}
}
}
