% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmr_utils.R
\name{calccaa}
\alias{calccaa}
\title{calccaa estimates the catch-at-age from teh predicted numbers-at-age}
\usage{
calccaa(pars, pnaa, M, sel, ages)
}
\arguments{
\item{pars}{the model parameter vector}

\item{pnaa}{the predicted numbers-at-age from the parameters}

\item{M}{the instantaneous natural mortality rate}

\item{sel}{the selectivity-at-age, which can be parameters}

\item{ages}{a vector of ages}
}
\value{
a matrix of predicted catch-at-age with yrs as rows and ages as cols
}
\description{
calccaa uses the Baranov catch equation to calculate the
predicted catch-at-age from teh predicted numbers-at-age (such as
calculates by calcnaa)
}
\examples{
print("wait on data-sets")
}
