% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calccaaC}
\alias{calccaaC}
\title{calccaaC calculates the predicted catch-at-age}
\usage{
calccaaC(pars, pnaa, M, sel, ages)
}
\arguments{
\item{pars}{a vector of log-transformed parameters}

\item{pnaa}{the predicted numbers-at-age in the population}

\item{M}{the natural mortality instantaneous rate}

\item{sel}{the selectivity at age}

\item{ages}{a vector identifying the ages being considered}
}
\description{
calccaaC calculate the expected catch-at-age for a
statistical catch-at-age model. Implemented in Rcpp.
}
