% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmr_utils.R
\name{calcnaa}
\alias{calcnaa}
\title{calcnaa estimates the predicted numbers-at-age from the parameters}
\usage{
calcnaa(pars, M, sel, yrs, ages)
}
\arguments{
\item{pars}{the model parameter vector}

\item{M}{the instantaneous natural mortality rate}

\item{sel}{the selectivity-at-age, which can be parameters}

\item{yrs}{a vector of yrs, used to label the rows}

\item{ages}{a vector of ages, used to label the columns}
}
\value{
a matrix of predicted numbers-at-age
}
\description{
calcnaa estimates the predicted numbers-at-age from the
parameters, along with the estimate of M, and selectivity-at-age
}
\examples{
print("wait on data-sets")
}
