% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcnaaC}
\alias{calcnaaC}
\title{calcnaaC calculates the predicted numbers-at-age and exploitable B}
\usage{
calcnaaC(pars, M, yrs, ages, owa)
}
\arguments{
\item{pars}{a vector of log-transformed parameters}

\item{M}{the natural mortality instantaneous rate}

\item{yrs}{is a vector of the years}

\item{ages}{a vector identifying the ages being considered}

\item{owa}{is a matrix of the observed weight-at-age}
}
\description{
calcnaaC calculates the expected numbers-at-age, the
exploitable-numbers-at-age, the exploitable biomass by year, and the
selectivity for a statistical catch-at-age model. Implemented in Rcpp.
}
