% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agestruct.R
\name{calcsel}
\alias{calcsel}
\title{calcsel estimates selectivity}
\usage{
calcsel(ageorlen, fishprop, typeselect, zerocomps = NULL)
}
\arguments{
\item{ageorlen}{a vector of ages or lengths against which to estimate the
selectivity}

\item{fishprop}{the fishery properties for a particular fleet from the
fishery list object out of getconstants}

\item{typeselect}{what sort of selectivity curve to use. Currently, the only
options are 'logistic' and 'domed' selectivity}

\item{zerocomps}{a vector of ages or sizes that are never selected. Identify
their indices in the outsel vector here. eg. if age 0 is never selected
then set zerocomps = 1, if ages 1-2 then zerocomps = c(1,2).}
}
\value{
a vector of selectivity for each ageorlen
}
\description{
calcsel estimates selectivity
}
\examples{
ages <- 0:25
pars <- c(0, 3, 1.0) # catchability + logistic parameters q, sel50, deltasel
cbind(ages,calcsel(ageorlen=ages,fishprop=pars,typeselect="logistic"))
pars <- c(0,5,9,16,23,-5,-2)
cbind(ages,calcsel(ageorlen=ages,fishprop=pars,typeselect="domed",
      zerocomp=1))
}
