% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{const}
\alias{const}
\title{const - conditioning data for an example age-structured stock}
\format{
A list containing 'biology', 'fishery', 'structure', and 'histcat':
\itemize{
\item {biology: M, surv, Linf, K, t0, growCV, Wta, Wtb, steep, age50M,
deltaM, B0, sigmaR, origsigR, and sigmaCE}
\item {fishery: sel50, deltasel, initdep, qc, nyrs, startyr, and endyr}
\item {structure: a vector of lengths from 0 - 78 in steps of 1, and a
vector of ages from 0 - 25}
\item {histcat: a vector of 35 historical catches fro 1980 - 2014}
}
}
\description{
const - A dataset containing the biology constants, the fishery
constants, and an historical catch dataset from 1976 - 2015, the values
in 'biology' relate to M, surv, Linf, K, t0, growCV, Wta, Wtb, steep,
age50M, deltaM, B0, sigmaR, origsigR, and sigmaCE. The values in 'fishery'
relate to sel50, deltasel, initdep, qc, nyrs, startyr, and endyr. Finally
there are the histcat historical catches from 1980 - 2014. This dataset
derives from the agestruct package
}
