% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agestruct.R
\name{datatemplateASPM}
\alias{datatemplateASPM}
\title{datatemplateASPM data template for a 2 fleet 1 region/stock fishery}
\usage{
datatemplateASPM(rundir, filename = "F2S1.csv", nfleet = 2)
}
\arguments{
\item{rundir}{directory in which to find the data file and run the analysis}

\item{filename}{the name of the data file to be produced, default=
'F21S.csv'.}

\item{nfleet}{the number of fleets for which data will be generated,
default = 2}
}
\value{
the function write a data file to rundir and returns the filename
}
\description{
datatemplateASPM write a data file for either a 1- or 2-fleet
stock
}
\examples{
\dontrun{
  rundir <- tempdir()
  dirExists(rundir,verbose=TRUE)
  datatemplateASPM(rundir,filename="Flt2Stock1.csv")
  dir(rundir)
}
}
