% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agestruct.R
\name{datatemplateIA}
\alias{datatemplateIA}
\title{datatemplateIA data template for a simple integrated stock assessment}
\usage{
datatemplateIA(rundir, filename = "F1S1-age.csv", nfleet = 1)
}
\arguments{
\item{rundir}{directory in which to find the data file and run the analysis}

\item{filename}{the name of the data file to be produced, default=
'F21S.csv'.}

\item{nfleet}{the number of fleets for which data will be generated,
default = 1; currently only 1 fleet is allowed for.}
}
\value{
the function write a data file to rundir and returns the filename
}
\description{
datatemplateIA writes a data file for either a 1- or 2-fleet
stock that also provide for age-composition data
}
\examples{
\dontrun{
  rundir <- tempdir()
  dirExists(rundir,verbose=TRUE)
  datatemplateIA(rundir,filename="Flt2Stock1-age.csv",nfleet=1)
  dir(rundir)
}
}
