% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmr-options.R
\docType{data}
\name{.cbPalette2}
\alias{.cbPalette2}
\title{The color-blind friendly palette with black}
\format{
An object of class \code{character} of length 8.
}
\usage{
.cbPalette2
}
\description{
see http://www.cookbook-r.com/Graphs/Colors_%28ggplot2%29/
}
\author{
DN Webber
}
\keyword{datasets}
