% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-catch.R
\name{.get_catch_df}
\alias{.get_catch_df}
\title{Get observed and predicted catch values}
\usage{
.get_catch_df(M)
}
\arguments{
\item{M}{List object(s) created by read_admb function}
}
\value{
dataframe of catch history (observed) and predicted values
}
\description{
Get observed and predicted catch values
}
\author{
SJD Martell, D'Arcy N. Webber
}
