% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-cpue.R
\name{.get_cpue_df}
\alias{.get_cpue_df}
\title{Get cpue or other indices}
\usage{
.get_cpue_df(M)
}
\arguments{
\item{M}{List object created by read_admb function}
}
\value{
dataframe of observed and predicted indices and residuals
}
\description{
Get cpue or other indices
}
\author{
SJD Martell, D'Arcy N. Webber
}
