% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-length-weight.R
\name{.get_length_weight_df}
\alias{.get_length_weight_df}
\title{Get length-weight data}
\usage{
.get_length_weight_df(M)
}
\arguments{
\item{M}{List object created by read_admb function}
}
\value{
dataframe of the length-weight relationship used in the model
}
\description{
Get length-weight data
}
\author{
DN Webber
}
