% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-natural-mortality.R
\name{.get_M_df}
\alias{.get_M_df}
\title{Get natural mortality data}
\usage{
.get_M_df(M)
}
\arguments{
\item{M}{list object created by read_admb function}
}
\description{
Get natural mortality data
}
\author{
J Ianelli, SJD Martell, DN Webber
}
