% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-maturity.R
\name{.get_maturity_df}
\alias{.get_maturity_df}
\title{Get maturity data}
\usage{
.get_maturity_df(M)
}
\arguments{
\item{M}{List object created by read_admb function}
}
\value{
dataframe of the length-weight relationship used in the model
}
\description{
Get maturity data
}
\author{
D'Arcy N. Webber
}
