% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-molt-prob.R
\name{.get_molt_prob_df}
\alias{.get_molt_prob_df}
\title{Get molting probability data}
\usage{
.get_molt_prob_df(M)
}
\arguments{
\item{M}{a list of lists created by the read_admb function}
}
\value{
dataframe of the molting probability data
}
\description{
Get molting probability data
}
\author{
D'Arcy N. Webber
}
