% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-numbers.R
\name{.get_numbers_df}
\alias{.get_numbers_df}
\title{Get numbers at length data}
\usage{
.get_numbers_df(M)
}
\arguments{
\item{M}{list object(s) created by read_admb function}
}
\value{
dataframe of numbers at length
}
\description{
Get numbers at length data
}
\author{
DN Webber
}
