% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-recruitment.R
\name{.get_recruitment_size_df}
\alias{.get_recruitment_size_df}
\title{Get recruitment size distribution data}
\usage{
.get_recruitment_size_df(M)
}
\arguments{
\item{M}{list object(s) created by read_admb function}
}
\value{
dataframe of recruitment size distribution
}
\description{
Get recruitment size distribution data
}
\author{
DN Webber
}
