% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-selectivity.R
\name{.get_selectivity_df}
\alias{.get_selectivity_df}
\title{Get the selectivity data}
\usage{
.get_selectivity_df(M)
}
\arguments{
\item{M}{list object created by read_admb function}
}
\value{
list of selectivities
}
\description{
Get the selectivity data
}
\author{
SJD Martell, D'Arcy N. Webber
}
