% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-size-comps.R
\name{.get_sizeComps_df}
\alias{.get_sizeComps_df}
\title{Get observed and predicted size composition values}
\usage{
.get_sizeComps_df(M)
}
\arguments{
\item{M}{a list of lists created by the read_admb function}
}
\value{
a list of observed and predicted size composition values
}
\description{
Get observed and predicted size composition values
}
\author{
SJD Martell, D'Arcy N. Webber
}
