% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\name{.onLoad}
\alias{.onLoad}
\title{On load hook}
\usage{
.onLoad(libname, pkgname)
}
\description{
This is a load hook that is called by R when the package is loaded. This
should not be exported
}
\author{
J Ianelli
}
