% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspm.r
\name{dynamicsF}
\alias{dynamicsF}
\title{dynamicsF describe the ASPM dynamics using instantaneous F}
\usage{
dynamicsF(
  pars,
  infish,
  inglb,
  inprops,
  year = "year",
  catch = "twl",
  cpue = "twlCE",
  waa = "waa",
  maa = "maa",
  sela = "twl",
  full = FALSE,
  diffC = TRUE,
  maxF = 4,
  ...
)
}
\arguments{
\item{pars}{the dynamics relies on many parameters sitting in the global
environment in particular ages, nages, maxage, M, fish, nyrs, and
maa, waa, sela, which are contained in props. 'pars' can contain either
two or three parameters. 1) is the log-transformed average unfished
recruitment, inR0. 2) is the variability around the index of relative
abundance (cpue or survey index) during the fitting process, and if is
present 3) if is present, is the catchability 'q',
which alternatively can be estimated using the closed form. 4) If
present this would be the log of the initail depletion.}

\item{infish}{the fish data.frame from readdata or built in dataset}

\item{inglb}{the glb data.frame from readdata or built in dataset}

\item{inprops}{the props data.frame from readdata or built in dataset}

\item{year}{the character name of time-step, default='year'}

\item{catch}{character name of catches, default='twl'}

\item{cpue}{character name of CPUE, default = 'twlCE'}

\item{waa}{the character name of the weight-at-age}

\item{maa}{the character name of the maturity-at-age}

\item{sela}{the character name of the selectivity-at-age}

\item{full}{should all outputs from dynamics be given. When fitting the
model, set this to FALSE. Once fitted, change this to TRUE to get all
the required outputs.}

\item{diffC}{should a catch penalty be used. default = TRUE}

\item{maxF}{search for F in optimize from 0 to maxF, default = 4.0}

\item{...}{any extra arguments needed}
}
\value{
a data.frame containing the fishery dynamics according to the input
parameter inR0. Includes Year, Catch, PredC, SpawnB, ExploitB, FullH,
CPUE, PredCE, Deplete, Recruit, FullF.
}
\description{
dynamicsF summarizes the dynamics of an Age-Structured
Production Model (ASPM). Fitting the ASPM entails estimating the unfished
recruitment level (R0), which is input as a parameter. In this case
fishing mortality is implemented as instantaneous rates rather that
as annual harvest rates. The maximize F = 4.0. It would be rare that
when using instantaneous fishing mortality rates that predicted catch
failed to equal observed catches. Even so, we have included the same
penalty term into the likelihood as used in dynamicsH, just to the sure.
}
\examples{
\dontrun{
data(westroughy)
fish <- westroughy$fish
glb <- westroughy$glb
props <- westroughy$props
pars <- c(7.0,0.3,-7.7) # logR0, sigCE, estimate avq
fishery <- dynamicsF(pars,fish,glb,props,full=TRUE,sela=TRUE) 
bestL <- optim(pars,dynamicsF,method="Nelder-Mead",infish=fish,inglb=glb,
               inprops=props,sela="twl",
               control=list(maxit = 1000,parscale = c(10,0.1,10)))
str(bestL)
fishery <- dynamicsF(bestL$par,fish,glb,props,full=TRUE,sela="twl")
print(round(fishery,4)) 
}
}
\seealso{
{
\link{dynamicsH}
}
}
