% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspm.r
\name{dynamicsH}
\alias{dynamicsH}
\title{dynamicsH describe the ASPM dynamics using annual harvest rates}
\usage{
dynamicsH(
  pars,
  infish,
  inglb,
  inprops,
  year = "year",
  catch = "twl",
  cpue = "twlCE",
  waa = "waa",
  maa = "maa",
  sela = "twl",
  full = FALSE,
  diffC = TRUE,
  ...
)
}
\arguments{
\item{pars}{the dynamics relies on many parameters sitting in the function's
environment, these are ages, nages, maxage, M, maa, waa, sela, fish,
and nyrs. 'pars' can contain either two to three log-transformed
parameters: 1) is the log of average unfished recruitment, inR0. 2) is
the variability around the index of relative abundance (cpue), not used
directly in the dynamics but rather in the estimation of the likelihoods
during the fitting process, 3) if is present, is the catchability 'q',
which alternatively can be estimated using the closed form. 4) If
present this would be the log of the initial depletion.}

\item{infish}{the fish data.frame from readdata or an internal dataset}

\item{inglb}{the glb data.frame from readdata or an internal dataset}

\item{inprops}{the props data.frame from readdata or an internal dataset}

\item{year}{the character name of time-step, default='year'}

\item{catch}{character name of catches, default='twl'}

\item{cpue}{character name of CPUE, default = 'twlCE'}

\item{waa}{the character name of the weight-at-age}

\item{maa}{the character name of the maturity-at-age}

\item{sela}{the character name of the selectivity-at-age,default='twl'}

\item{full}{should all outputs from dynamics be given. When fitting the
model, set this to FALSE. Once fitted, change this to TRUE to get all
the required outputs.}

\item{diffC}{should a catch penalty be used. default = TRUE}

\item{...}{any extra arguments needed}
}
\value{
a data.frame containing the fishery dynamics according to the input
parameter inR0. In particular it includes the Catch and PredC, and the
CPUE and PredCE, which can be used in a maximum likelihood context.
}
\description{
dynamicsH summarizes the dynamics of the Age-Structured
Production Model (ASPM) in which the catches are predicted from annual
harvest rates. Fitting the ASPM entails estimating the unfished
recruitment level (R0), which is input as a parameter. There may be
other parameters as described in the pars section. Methematically, it is
possible to have harvest rates > 1.0, which is biological nonsense. So,
the code includes an if statement that constrains the harvest rate to
< 0.975 (which would be ridiculously high). Because of this we also \
include a penalty, absdiffC, to ensure that the predicted catches = the
observed catches.
}
\examples{
data(westroughy)
fish <- westroughy$fish
glb <- westroughy$glb
props <- westroughy$props
pars=c(7.064324,-1.257487,-7.694248)
bestL <- optim(pars,dynamicsH,method="Nelder-Mead",infish=fish,inglb=glb,
               inprops=props,control=list(maxit=1000,parscale = c(10,1,10)))
str(bestL)
codeutils::outfit(bestL,digits=6)
out <- dynamicsH(bestL$par,infish=fish,inglb=glb,inprops=props,full=TRUE)
print(round(out$fishery,4))
# pars=c(7.064324,-1.257487,-7.694248);infish=fish;inglb=glb;
# inprops=props;full=TRUE; waa="waa";maa="maa";sela="twl"
}
\seealso{
{
\link{dynamicsF}
}
}
