% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspm.r
\name{dynF}
\alias{dynF}
\title{dynF describe the ASPM dynamics}
\usage{
dynF(
  pars,
  infish,
  inglb,
  inprops,
  waa = "waa",
  maa = "maa",
  sela = "twl",
  year = "year",
  catch = "twl",
  cpue = "twlCE",
  full = FALSE,
  reps = 5
)
}
\arguments{
\item{pars}{the dynamics relies on many parameters sitting in the global
environment in particular ages, nages, maxage, M, fish, nyrs, and
maa, waa, sela, which are contained in props. 'pars' can contain either
two or three parameters. 1) is the log-transformed average unfished
recruitment, inR0. 2) is the variability around the index of relative
abundance (cpue or survey index) during the fitting process, and if is
present 3) if is present, is the catchability 'q',
which alternatively can be estimated using the closed form. 4) If
present this would be the log of the initail depletion.}

\item{infish}{the fish data.frame from readdata or built in dataset}

\item{inglb}{the glb data.frame from readdata or built in dataset}

\item{inprops}{the props data.frame from readdata or built in dataset}

\item{waa}{the character name of the weight-at-age}

\item{maa}{the character name of the maturity-at-age}

\item{sela}{the character name of the selectivity-at-age,default='twl'}

\item{year}{the character name of the year column in fish, default='year'}

\item{catch}{the character name of the catch column in fish, default='twl'}

\item{cpue}{the character name of the cpue column in fish, default='twlCE'}

\item{full}{should all outputs from dynamics be given. When fitting the
model, set this to FALSE. Once fitted, change this to TRUE to get all
the required outputs.}

\item{reps}{how many loops within the findF function to find each F}
}
\value{
a data.frame containing the fishery dynamics according to the input
parameter inR0. Includes Year, Catch, PredC, SpawnB, ExploitB, FullH,
CPUE, PredCE, Deplete, Recruit, FullF.
}
\description{
dynF summarizes the dynamics of an Age-Structured
Production Model (ASPM). Fitting the ASPM entails estimating the unfished
recruitment level (R0), which is input as a parameter. In this case
fishing mortality is implemented as instantaneous rates rather that
as annual harvest rates. The maximize F = 4.0. Once again we have
included a penalty on catches to ensure that predC = ObsC.
}
\examples{
\dontrun{
data(westroughy)
fish <- westroughy$fish
glb <- westroughy$glb
props <- westroughy$props
pars <- c(7.1,-1,-7.7) # logR0, sigCE, estimate avq
fishery <- dynF(pars,fish,glb,props) 
pars <- c(7.1,-1,-7.7)
bestL <- optim(pars,dynF,method="Nelder-Mead",infish=fish,inglb=glb,
               inprops=props,control=list(maxit=1000,parscale = c(10,1,10)))
str(bestL)
codeutils::outfit(bestL,digits=6)
out <- dynF(bestL$par,fish,glb,props,full=TRUE)
print(round(out$fishery,4)) 
}
# pars=pars <- c(7.0,-0.7,-6.7) ;infish=fish;inglb=glb;inprops=props
# waa="waa";maa="maa";sela="sela"; full=TRUE; reps=5
}
\seealso{
{
\link{dynamicsF}, \link{dynamicsF}, \link{findF}
}
}
