% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{F2robustness}
\alias{F2robustness}
\title{F2robustness are robustness outputs for 2 Fleet 1 region fishery}
\format{
A data.frame of 15 columns and 200 rows
}
\usage{
F2robustness
}
\description{
F2robustness a data.frame containing the results of running
a robustness test on the two fleet one region/stock fishery example.
This uses the dynF2 function to describe the stock dynamics using the
two fishing gears with different selectivity and instantaneous
fishing mortality rates. It constitutes a 200 x 15 data.frame with
columns iLnR0, iqc1, iqc2, -veLL, LnR0, Lqc1, Lqc2, gradR0, gradqc1,
gradqc2, R0, qc1, qc2, pardist, and Iters. Where pardist is the
Euclidean distance between the initial parameters and the original
optimum. Each nominal parameter can be changed by up to 10 percent. In
practice LnR0 was changed by up to 4.99 percent, Lnq1 by up to 8.90
percent, and Lnq2 buy up to 10.32 percent. The results are sorted from
the smallest to the largest -veLL, although the biggest difference is
only 0.00082 likelihood units. This is a robust result.
}
\examples{
 data(F2robustness)
 str(F2robustness)
 countLL <- table((round(F2robustness[,"-veLL"],6)))
 countLL
 range(F2robustness[,"-veLL"]) # should be -11.35589 -11.35562
}
\keyword{datasets}
