% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{findFs}
\alias{findFs}
\title{findFs uses an iterative strategy for finding each year's F values}
\usage{
findFs(cyr, Nyr, sel, wata, M, reps = 8)
}
\arguments{
\item{cyr}{a vector of known catches in a given year for nfleets}

\item{Nyr}{the numbers at size at the start of the given year or end of the
year before}

\item{sel}{a matrix of the selectivity of the fishing gears}

\item{wata}{the weight-at-age, might be waa vector / 1000.0}

\item{M}{the instantaneous natural mortality rate}

\item{reps}{how many internal loops to use finding each F, default = 6}
}
\value{
the fully selected fishing mortality rates
}
\description{
findFs is used when searching for the instantaneous F that will
produce the required catch in a given year in an age-structured model
when there are multiple gears, with different selectivity being used. It
uses a method modified from that described in Methot and Wetzel
(2013, p7). This arrangement is designed solely for annual fisheries.
By including a 'season' term and dividing each estimate of fFyr by
the fraction of a year in each 'season', this adjustment routine could
be applied to seasonal fisheries.
}
\examples{
  require(fmr)
 library(codeutils)
 library(hplot)
 data("westroughy")
 fish <- westroughy$fish
 glb <- westroughy$glb
 props <- westroughy$props
 pars <- c(7,-0.4,-6.7)  
 bestFD <- fitASPM(initpar=pars,minfun=dynF,infish=fish,inglb=glb,
                   inprops=props,gradtol=1e-05,stepmax=0.1,steptol=1e-07,
                   hessian=TRUE,reps=6)
 outfit(bestFD,digits=7,title="Instantaneous Rates 2",
        parnames=c("LnR0","Ln(sigCE)","Ln(q)")) 
 # cyr=obsC[1];Nyr=Nt[,yr];sel=sel;aaw=aaw;M=M;reps=reps  
}
\references{
Methot, R.D. and C.R. Wetzel (2013) Stock synthesis: A biological
and statistical framework for fish stock assessment and fishery
management. Supplementary Material: Appendix A: Technical Description of
the Stock Synthesis assessment program \emph{Fisherie Research} 142:
86-99. http://dx.doi.org/10.1016/j.fishres.2012.10.012
}
