% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{fish}
\alias{fish}
\title{fish fishery data for the plaice example from Beverton and Holt, 1957}
\format{
A data.frame with four columns
\describe{
\item{year}{ 9 fishing years from 1929 to 1937, listed as 29 - 37}
\item{effort}{ the annual effort}
\item{catch}{ the annual catch of plaice}
\item{obsce}{ the cpue calculated by dividing the catch by the effort}
}
}
\usage{
fish
}
\description{
A dataset containing the fish data.frame, for use in fmr
}
\examples{
data(fish)
fish
}
\references{
Beverton, R.J.H. and S.J. Holt (1957) \emph{On the dynamics of
exploited fish populations.} U.K. Ministry of Agriculture and Fisheries,
Fisheries Investigations (Series 2), \emph{19}: 1-533.
}
\keyword{datasets}
