% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fishing_zone}
\alias{fishing_zone}
\title{Fishing zones}
\format{
\subsection{\code{fishing_zone}}{

A tibble
}
}
\usage{
fishing_zone
}
\description{
A vector containing variable names as delivered in the the FM-API's and
the corresponding names used in {fmr} if arguement std is set to TRUE
}
\keyword{datasets}
