% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{fitlbm}
\alias{fitlbm}
\title{fitsdm fits a size-based model for example abalone data}
\usage{
fitlbm(
  pardat,
  likefunk,
  funk,
  glb,
  fish,
  biol,
  G,
  const,
  omega,
  sizecomp,
  both = TRUE,
  tol = 1e-08,
  initH = 0
)
}
\arguments{
\item{pardat}{a two column matrix of the log-trasnformed parameter vector
combined with another column of phases, either 1 or 0 determiing
whether the parameter will be estimated or not}

\item{likefunk}{a function for estimating the negative log-likelihood -veLL}

\item{funk}{a function used to calcualte the stock dynamics}

\item{glb}{the globals object see str(glb)}

\item{fish}{the fisheries dependent datam year, LML, catch and cpue}

\item{biol}{the biol matrix from makebiol}

\item{G}{the growth transition matrix from makebiol}

\item{const}{a vector of biological constants used to set weight-, maturity-,
emergence-, and growth-at-size. Emergence is of post-larval juveniles
from cryptic habitat.}

\item{omega}{a vector of relative weights applied to each data set in an
effort to balance their relative contribution to the total -veLL}

\item{sizecomp}{the size composition data from commercial catches, usually
after filtering for some minimum sample size number.}

\item{both}{default = TRUE Should the detailed components of the -veLL be
returned, - used when summarizing the optimum fit, or, the default,
should only the -veLL be returned - used when fitting the model}

\item{tol}{the tolerance used when using the numerical methods with optim
and nlminb}

\item{initH}{an annual harvest rate used to set an initial depletion,
default = 0}
}
\value{
Either just the -veLL, the default, or a more detailed vector of
the components of the -veLL and other constants used in the model fit.
}
\description{
fitsbm is a wrapper function that sets up the input requirements
for optim and, optionally, nlminb so it allows a dual model fit first
with optim, which is robust but generally slow, followed by nlminb which
is faster but, usually only if it starts near the solution.
}
\examples{
\dontrun{
 require(fmr)
 data(absaudata)
 list2env(absaudata,envir=.GlobalEnv)
 biolout <- makeabiol(fish=fish,const=const,glb=glb)
 out <- abdynamicsH(pin=pindat[,1],fish=fish,biol=biol,G=G,const=const,
                    glb=glb)
 picksc=which(colSums(sizecomp) > 100)
 finalcomp <- sizecomp[,picksc]
 omega=c(1,1e-03)
 fitout <- fitlbm(pardat=pindat,likefunk=negLL,funk=abdynamicsH,glb=glb,
               fish=fish,biol=biol,G=G,const=const,omega=omega,
               sizecomp=finalcomp)
 out <- abdynamicsH(pin=fitout$optpar,fish=fish,biol=biol,G=G,const=const,
                    glb=glb)
 plotSBMdynamics(rundir="",dyn=out$dyn,fish=fish,pars=exp(fitout$optpar),
                 glb,console = TRUE) 
 pindat <- cbind(fitout$optpar,rep(1,length(fitout$optpar)))
 print(pindat)
}
}
