% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice.R
\name{fm_choice}
\alias{fm_choice}
\title{choice table (choiceD)}
\usage{
fm_choice(key, std = TRUE)
}
\arguments{
\item{key}{your FM API key}

\item{std}{boolean, if TRUE (default) returns shorter names}
}
\value{
a tibble
}
\description{
choice table (choiceD)
}
