% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_aux_tables.R
\name{fm_coordinateD}
\alias{fm_coordinateD}
\title{coordinate table (coordinateD)}
\usage{
fm_coordinateD(key)
}
\arguments{
\item{key}{your FM API key}
}
\value{
a tibble
}
\description{
coordinate table (coordinateD)
}
