% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadump.R
\name{fm_datadump}
\alias{fm_datadump}
\title{Fisheries manager datadump}
\usage{
fm_datadump(key, pth = "dump")
}
\arguments{
\item{key}{your FM API key}

\item{pth}{vector specifying path to were the data is to be stored}
}
\description{
Dumps all available API data from Fisheries Manager as a R rds-file
to the path specified.
}
