% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadump.R
\name{fm_dump}
\alias{fm_dump}
\title{Dump a FM table as rds}
\usage{
fm_dump(table, key, dump_pth)
}
\arguments{
\item{table}{FM table name}

\item{key}{your FM API key}

\item{dump_pth}{pth to your}
}
\description{
Dump a FM table as rds
}
