% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_table.R
\name{fm_gear}
\alias{fm_gear}
\title{Gear table}
\usage{
fm_gear(key, std = TRUE, trim = TRUE)
}
\arguments{
\item{key}{string, your API key to FTM}

\item{std}{boolean (default TRUE) determining naming convention}

\item{trim}{boolean (default TRUE), if only "essential" variables}
}
\value{
a tibble
}
\description{
Gear table
}
