% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{fm_grid_calendar}
\alias{fm_grid_calendar}
\title{Creates a monthly grid}
\usage{
fm_grid_calendar(d)
}
\arguments{
\item{d}{a tibble containing varible date}
}
\value{
a tibble with added parameter .wday, .week, .month and .day
}
\description{
Creates a monthly grid
}
