% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landing.R
\name{fm_landingF}
\alias{fm_landingF}
\title{landingF data-warehouse table}
\usage{
fm_landingF(key)
}
\arguments{
\item{key}{your FM API key}
}
\value{
a tibble
}
\description{
A wide table consisting elements from the survey_table, survey_item table and
survey_item_dtl table
}
