% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_table.R
\name{fm_landings_site}
\alias{fm_landings_site}
\title{Landing sites}
\usage{
fm_landings_site(key)
}
\arguments{
\item{key}{string, your API key to FTM}

\item{std}{boolean (default TRUE) determining naming convention}

\item{trim}{boolean (default TRUE), if only "essential" variables}
}
\value{
A tibble
}
\description{
Landing sites
}
