% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landing.R
\name{fm_landingV}
\alias{fm_landingV}
\title{landingV data-warehouse table}
\usage{
fm_landingV(key)
}
\arguments{
\item{key}{your FM API key}
}
\value{
a tibble
}
\description{
A wide table consisting elements from the survey_table, survey_item table and
survey_item_dtl table as well as variables from auxillary tables.
}
