% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_aux_tables.R
\name{fm_licenseF}
\alias{fm_licenseF}
\title{license fact table (licenseF)}
\usage{
fm_licenseF(key)
}
\arguments{
\item{key}{your FM API key}
}
\value{
a tibble
}
\description{
license fact table (licenseF)
}
