% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_aux_tables.R, R/geo.R
\name{fm_location}
\alias{fm_location}
\title{location table (locationD)}
\usage{
fm_location(key)

fm_location(key)
}
\arguments{
\item{key}{your FT API key}
}
\value{
a tibble

A tibble
}
\description{
location table (locationD)

Location table
}
