% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{fm_location_knr}
\alias{fm_location_knr}
\title{Coordinates of fishing zones}
\usage{
fm_location_knr(key)
}
\arguments{
\item{key}{you FT API key}
}
\value{
a tibble
}
\description{
Specific for St. Kitt's and Nevis
}
