% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{fm_read_country}
\alias{fm_read_country}
\title{Get GADM shapefile for a country}
\usage{
fm_read_country(country_code = "KNA")
}
\arguments{
\item{country_code}{country 3 letter ISO code}
}
\value{
a tiblle
}
\description{
Get GADM shapefile for a country
}
