% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_tables.R
\name{fm_sidF}
\alias{fm_sidF}
\title{The survey_item_dtl variables in landingF}
\usage{
fm_sidF(key)
}
\arguments{
\item{key}{your FM API key}
}
\value{
a tibble
}
\description{
The survey_item_dtl variables in landingF
}
