% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_tables.R
\name{fm_siF}
\alias{fm_siF}
\title{The survey_item variables in landingF}
\usage{
fm_siF(key)
}
\arguments{
\item{key}{your FM API key}
}
\value{
a tibble
}
\description{
The survey_item variables in landingF
}
