% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site.R
\name{fm_site}
\alias{fm_site}
\title{Sites table}
\usage{
fm_site(key, std = TRUE, trim = TRUE)
}
\arguments{
\item{key}{string, your API key to FTM}

\item{std}{boolean (default TRUE) determining naming convention}

\item{trim}{boolean (default TRUE), only return "essential" variables}
}
\value{
A tibble
}
\description{
In a catch-sampling cse this constitutes a landings site table
}
