% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{fm_species}
\alias{fm_species}
\title{species table (specieD)}
\usage{
fm_species(key, std = TRUE, trim = TRUE)
}
\arguments{
\item{key}{your FM API key}

\item{std}{boolean (default TRUE) determining naming convention}

\item{trim}{boolean (default TRUE), if only "essential" variables}
}
\value{
a tibble
}
\description{
species table (specieD)
}
