% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_tables.R
\name{fm_survey_item}
\alias{fm_survey_item}
\title{survey_item table}
\usage{
fm_survey_item(key, std = TRUE, trim = TRUE)
}
\arguments{
\item{key}{your FM API key}

\item{std}{boolean}

\item{trim}{boolean}
}
\value{
a tibble
}
\description{
survey_item table
}
