% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_aux_tables.R
\name{fm_surveydateD}
\alias{fm_surveydateD}
\title{surveydate table (surveydateD)}
\usage{
fm_surveydateD(key)
}
\arguments{
\item{key}{your FM API key}
}
\value{
a tibble
}
\description{
surveydate table (surveydateD)
}
