% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyitem.R
\name{fm_surveyitem}
\alias{fm_surveyitem}
\title{The surveyitem table}
\format{
A data frame with these variables if using default arguements:
\describe{
\item{\code{xxx}}{yyy}
}
}
\usage{
fm_surveyitem(key, std = TRUE, trim = TRUE, remove_empty = TRUE)
}
\arguments{
\item{key}{your FM API key}

\item{std}{boolean, if TRUE (default) variables are renamed according to
{fmr} convention}

\item{trim}{boolean, if TRUE (default) only essential (standard) variables
are returned. Only active if arguement std is set to TRUE.}

\item{remove_empty}{boolean, if TRUE (default) remove variables whose values
are all NA's}
}
\value{
a tibble
}
\description{
The function returns information about each ...
}
