% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyitemdtl.R
\name{fm_surveyitemdtl_api}
\alias{fm_surveyitemdtl_api}
\title{The surveyitemdtl api table}
\usage{
fm_surveyitemdtl_api(key)
}
\arguments{
\item{key}{your FM API key}
}
\value{
a tibble
}
\description{
The function gets the data from the FM api and ensure correct variable types.
Variable names and order as as it comes from the API. The user should
normally use the fm_surveyitem function.
}
