% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyitem.R
\name{fm_trip}
\alias{fm_trip}
\title{A trip table}
\usage{
fm_trip(key)
}
\arguments{
\item{key}{your FM API key}
}
\value{
a tibble
}
\description{
A trip table
}
\note{
For now same as fm_surveytable with default arguments. In future will
most likely need a filter using one of the _id, _class or _category
}
