% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_aux_tables.R
\name{fm_vessel}
\alias{fm_vessel}
\title{Vessel table}
\usage{
fm_vessel(key)
}
\arguments{
\item{key}{your FM API}
}
\value{
a tibble
}
\description{
For the time being simple, only 3 variables returned
}
