% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_aux_tables.R
\name{fm_vessel_class}
\alias{fm_vessel_class}
\title{Vessel class table (vesselclassD)}
\usage{
fm_vessel_class(key)
}
\arguments{
\item{key}{your FM API}
}
\value{
a tibble
}
\description{
Vessel class table (vesselclassD)
}
