% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vessel.R
\name{fm_vesselclassD}
\alias{fm_vesselclassD}
\title{Vessel class table (vesselclassD)}
\usage{
fm_vesselclassD(key)
}
\arguments{
\item{key}{your FM API}
}
\value{
a tibble
}
\description{
Vessel class table (vesselclassD)
}
