% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmr.R
\docType{package}
\name{fmr}
\alias{fmr-package}
\alias{fmr}
\title{fmr R functions for Fisheries Modelling using R}
\description{
The fmr package provides R functions and example data sets for
use with the new book: (\emph{Fisheries Modelling Using R}), that will
only be available as a GitBook at (\emph{haddonm.github.io/fmr}). This
new book is designed to provide chapters on somewhat more advanced topics
than those in Haddon (2021).  The scr files will be added later
# #' need an 'at' here useDynLib fmr, .registration = TRUE
}
\references{
Haddon, M. (2021) Using R for Modelling and Quantitative
Methods in Fisheries, CRC Press / Chapman & Hall/ Boca Raton 337p.
ISBN: 9780367469894
}
\keyword{internal}
