% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{fournarch82}
\alias{fournarch82}
\title{fournarch82 the observed catch-at-age for plaice}
\format{
A 20 row x 10 column  matrix purporting to be the number-at-age in a
simulated population across 20 years for 10 ages, 4 - 13.
}
\usage{
fournarch82
}
\description{
Table 1 from Fournier and Archibald, 1982, p1200 containing the
actual numbers-at-age in a simulated population.
}
\examples{
data(fournarch82)
fournarch82
rowfreqboot(x=fournarch82,n=250)
}
\references{
Fournier, D. and C.P. Archibald (1982) A general theory for
analyzing catch at age data. \emph{Canadian Journal of Fisheries and
Aquatic Sciences} \emph{39}: 1195-1207.
}
\keyword{datasets}
