% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{francis92}
\alias{francis92}
\title{francis92 Three data objects suitable for use with an ASPM}
\format{
A list of three objects
\describe{
\item{fish}{ a data.frame containing year, catch (tonnes), index (trawl
survey index), and Coefficient of Variation.}
\item{glb}{ a list of global variables including maxage, M, parameters for
growth, weight-at-age, maturity-at-age, steepness, R0, selectivity,
resilience, number of ages, ages, number of years of catch data,
and the name of the species. }
\item{props}{ a data.frame of age, laa, waa, maa, and sela, that is the
length-, weight, maturity, and selectivity-at-age}
}
}
\usage{
francis92
}
\description{
A dataset containing the fish data.frame, the glb list, and the
props data.frame set up ready for use with an ASPM. These are data
obtained from Francis 1992 concerning orange roughy from New Zealand's
Chatham Rise. Francis' 1992 paper is well known as an important
discussion of risk analysis when providing fisheries management advice.
It was one of the first published uses of an age-structured production
model. It is even better known as providing a clear exposition of what
has become a standard re-parameterization of the Beverton-Holt stock-
recruitment model. Francis demonstrates how the B-H model can be
re-defined using steepness and unfished recruitment level. The full
derivation is provided in Haddon (2021, p172).
}
\examples{
data(francis92)
str(francis92)

}
\references{
Francis, R.I.C.C. (1992) Use of Risk Analysis to Assess Fishery
Management Strategies: A Case Study using Orange Roughy
(\emph{Hoplostethus atlanticus}) on the Chatham Rise, New Zealand.
\emph{Canadian Journal of Fisheries and Aquatic Sciences} \emph{49}:
922-930.

Haddon, M. (2021) \emph{Using R for Modelling and Quantitative
Methods in Fisheries} CRC Press. Chapman & Hall, Boca Raton. 337p. ISBN:
978-0-367-46988-7 see also https://haddonm.github.io/URMQMF/index.html.
}
\keyword{datasets}
