% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_connect.R
\name{ft_connect}
\alias{ft_connect}
\title{Connect to Fisheries Technology Database}
\usage{
ft_connect(username, password)
}
\arguments{
\item{username}{your username}

\item{password}{your password}
}
\value{
a database connection
}
\description{
This function is the mother of all mothers that all other ft_ functions
rely on. If not you are doomed.
}
