% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_connect.R
\name{ft_tbl}
\alias{ft_tbl}
\title{Access a table}
\usage{
ft_tbl(con, tbl)
}
\arguments{
\item{con}{A oracle connection}

\item{tbl}{A string containing schema and table, separated by a dot}
}
\description{
Access a table
}
