% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{getabSBMinit}
\alias{getabSBMinit}
\title{getabSBMinit initiates a single sex abalone Size-Based Model}
\usage{
getabSBMinit(R0, glb, biol)
}
\arguments{
\item{R0}{the nominal scale estimated unfished recruitment}

\item{glb}{the globals object containing natM and the midpts}

\item{biol}{the biol list containing growth, maturity and weight-at-length}
}
\value{
a list of B0 the unfished spawning biomass, and N0 the unfished
numbers-at-size
}
\description{
getabSBMinit estimates the B0, N0, and NEbeg (initial start of
year exploitable numbers-at-size), for a single sex abalone stock. It
requires the unfished recruitment R0, the globals object, glb, the
biol object of the species being modelled.
}
\examples{
data(absaudata)
abstk <- makeabstock(fish=absaudata$fish,const=absaudata$const,
                     glb=absaudata$glb,pin=absaudata$pindat)
initSBM <- getabSBMinit(exp(absaudata$pindat["LnR0",1]),glb=absaudata$glb,
                      biol=abstk$biol)
print(initSBM$B0)
print(round(initSBM$N0))
}
