% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{getabSBMproduction}
\alias{getabSBMproduction}
\title{getabSBMproduction calculates productivity from optimum parameters}
\usage{
getabSBMproduction(
  LML,
  optpar,
  glb,
  biol,
  G,
  nyrs = 100,
  H = seq(0.005, 0.4, 0.005)
)
}
\arguments{
\item{LML}{what legal minimum length to use?}

\item{optpar}{the optimum parameters (in log-space) from a model fit}

\item{glb}{the globals object}

\item{biol}{the biological properties of the population}

\item{G}{the growth transition matrix generated using the optimum parameters}

\item{nyrs}{how many years should one project the model when searching for
a new equilibrium when being fished at each harvest rate, default=100}

\item{H}{a vector of trial harvest rates, default = seq(0.005,0.4,0.005)}
}
\value{
a matrix of H, matureB, exploitableB, yield, depletion, and cpue
}
\description{
getabSBMproduction uses the optimum parameters and G to
describe the productivity characteristics of a population at a given LML.
It does this by fishing the equilibrium population to a new equilibrium
under and array of constant harvest rates to determine the sustainable
yields obtainable at each harvest rate used.
}
\examples{
print("wait on data sets")
#   LML=140;optpar=pindat[,1];glb=glb;biol=biol;G=G;nyrs=100;
#  H=seq(0.005,0.4,0.005)
}
