% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{getConst}
\alias{getConst}
\title{getConst extracts 'nb' numbers from a line of text}
\usage{
getConst(inline, nb, index = 2)
}
\arguments{
\item{inline}{text line to be parsed, usually obtained using readLines}

\item{nb}{the number of numbers to extract}

\item{index}{which non-empty object to begin extracting from?}
}
\value{
a vector of length 'nb'
}
\description{
getConst parses a line of text and extracts 'nb' pieces of
text as numbers. If there are missing values this will stop the
function with a warning. If RStudio drops into debug, toggle 'message
only' under Debug/on Error to avoid that, and put NA where the data
are missing
}
\examples{
  inline <- "name , 31,32,33"
  getConst(inline=inline,nb=3,index=2)
  inline <- "anothername , 31,32,NA"
  getConst(inline=inline,nb=3,index=2)
  inline <- "finalname , 31,32,"  # missing values not allowed   
 # getConst(inline=inline,nb=3,index=2) # unhash and try to see error
}
