% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmr_utils.R
\name{getq}
\alias{getq}
\title{getq calculates catchability from cpue and predicted exploitable biomass}
\usage{
getq(cpue, exB)
}
\arguments{
\item{cpue}{the vector of observed cpue}

\item{exB}{the output vector of predicted exploitable biomass}
}
\value{
the catchability as a scaler
}
\description{
getq uses a closed form version of the catchability that uses
the observed cpue and the predicted exploitable biomass to estimate the
scaling factor between erxploitable biomass and cpue.
}
\examples{
cpue <- c(0.369,0.254,0.407,0.357,0.326,0.284,0.403,0.297,0.397)
exB <- c(3712,3251,3810,4499,4500,3977,3972,3220,3386) 
getq(cpue,exB)  # should be 8.964018e-05 
}
