% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{getrmse}
\alias{getrmse}
\title{getrmse calculates the rmse of the input 'cpue' series}
\usage{
getrmse(indat, invar = "cpue", inyr = "year", natlog = TRUE)
}
\arguments{
\item{indat}{the matrix, spmdat, or data.frame containing both a 'year'
column and an invar column (default to 'cpue')}

\item{invar}{the column name of the variable whose rmse is wanted; defaults
to 'cpue'}

\item{inyr}{the column name that points to the 'year' name}

\item{natlog}{should we estimate the rmse of log-transformed data. default
= TRUE because we are usually dealing with cpue.}
}
\value{
a list of the rmse and the loess predicted values of the invar
for each year in the time-series
}
\description{
getrmse calculates the root mean square error (rmse) of the
input invar series (defaults to 'cpue') against an input 'year' time
series. This is primarily designed to generate an alternative estimate
of the intrinsic variability of a cpue time-series to that which may be
obtained from a cpue standardization. This enables one to use the Francis
2011 weighting scheme for indices of relative abundance, which he derived
from Clarke and Hare, 2006.
}
\examples{
year <- 1986:1994
cpue <- c(1.2006,1.3547,1.0585,1.0846,0.9738,1.0437,0.7759,1.0532,1.284)
dat <- as.matrix(cbind(year,cpue))
getrmse(dat,invar="cpue")  # should be 0.08265127
getrmse(dat,invar="cpue")$rmse
}
\references{
Clark, W.G., and S.R. Hare (2006) Assessment and management of
Pacific halibut: data, methods, and policy. \emph{Scientific Report 83.
International Pacific Halibut Commission}, Seattle, Washington. 104p.

Francis, R.I.C.C. (2011) Data weighting in statistical fisheries
stock assessment models. \emph{Canadian Journal of Fisheries and Aquatic
Sciences} \bold{68}: 1124-1138.
}
