% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{getSBMinit}
\alias{getSBMinit}
\title{getSBMinit initiates a Size-Based Model}
\usage{
getSBMinit(inR0, glb, biol, sex)
}
\arguments{
\item{inR0}{the nominal scale estimated unfished recruitment}

\item{glb}{the globals object containing natM and the midpts}

\item{biol}{the biol list containing growth, maturity and weight-at-length}

\item{sex}{which gender is being analysed? Either 1, females or 2, males}
}
\value{
a list of B0 the unfished spawning biomass, and N0 the unfished
numbers-at-size
}
\description{
getSBMinit estimates the B0, N0, and NEbeg (initial start of
year exploitable numbers-at-size), all by gender. It requires the
unfished recruitment R0, the globals object, glb, the biol object, and
the sex being modelled.
}
\examples{
print("Wait on datasets")
}
