% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{getsingleNum}
\alias{getsingleNum}
\title{getsingleNum find a line of text and extracts a single number}
\usage{
getsingleNum(varname, intxt)
}
\arguments{
\item{varname}{the name of the variable to get from intxt}

\item{intxt}{text to be parsed, usually obtained using readLines}
}
\value{
a single number or, if no value is in the data file a NULL
}
\description{
getsingleNum uses grep to find an input line. If the variable
being searched for fails then NULL is returned
}
\examples{
\dontrun{
 txtlines <- c("replicates, 100","Some_other_text, 52")
 getsingleNum("replicates",txtlines)
 getsingleNum("eeplicates",txtlines)
 getsingleNum("other",txtlines)
}
}
