% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{getStr}
\alias{getStr}
\title{getStr obtains a string from an input text line}
\usage{
getStr(inline, nb, index = 2)
}
\arguments{
\item{inline}{input text line with components separated by ','}

\item{nb}{number of parts to return}

\item{index}{which part to start returning to the program, default=2}
}
\value{
a vector of character string(s)
}
\description{
getStr obtains a string from an input text line in
which any parts are separated by ','. Then, after ignoring the
first component, assumed to be a label, it returns the first
nb parts.
}
\examples{
  txt <- "runlabel, development_run, label for this particular run"
  getStr(txt,1)
}
