% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{getvcov}
\alias{getvcov}
\title{getvcov calculates the hessian and the variance covariance matrix}
\usage{
getvcov(
  pardat,
  likefunk,
  dynfunk,
  glb,
  fish,
  biol,
  G,
  const,
  omega,
  sizecomp,
  initH = 0,
  num95 = 3,
  verbose = TRUE
)
}
\arguments{
\item{pardat}{the matrix of optimal parameters and which are to fitted}

\item{likefunk}{what function is used to calculate the -veLL}

\item{dynfunk}{what function estimates the stock dynamics}

\item{glb}{the global constants}

\item{fish}{the fishery data in two matrices. The first contains the
year, the LML, catches, and the cpue, and the second contains the
numbers-at-size in the catch, just for those years with observations.}

\item{biol}{the matrix of maturity, WtL, emergence, and Maturity x Weight}

\item{G}{the growth transition matrix from makebiol}

\item{const}{contains biological constants, including those for growth}

\item{omega}{the individual allocated weights for each data source}

\item{sizecomp}{the cleaned size-composition data}

\item{initH}{default=0. the initial harvest rate to achieve the initial
deplation. If that is 1.0 then initH = 0 works fine.}

\item{num95}{how many of the parameters should have 95 percent CI generated.
default=3, which, if none are held constant, will give all the non-recdev
parameters}

\item{verbose}{default = TRUE, which writes 95 percent CI to the console}
}
\value{
a list of the variance-covariance matrix, the standard errors of
all parameters, the correlation matrix between parameters, and a
matrix of parameter confidence intervals.
}
\description{
getvcov uses optim and the optimal parameter set to estimate
the hessian matrix, which it then inverts to estimate the variance-
covariance matrix. The standard errors of the parameters are found
on the diagonal. If verbose is TRUE, the default, then approximate 95
percent confidence intervals for the first 5 parameters are written to
the console. They are approximate because the number of observations is
set arbitrarily to 50.
}
\examples{
print("wait on example data sets")
#  pardat=pindat; likefunk=negLL;dynfunk=abdynamicsH;glb=glb;fish=fish
#  const=const; omega=omega; sizecomp=finalcomp;initH=0;verbose=TRUE;
#  num95=4
}
