% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agestruct.R
\name{getvect}
\alias{getvect}
\title{getvect extracts 'n' numbers from an identified line of text}
\usage{
getvect(varname, intxt, n)
}
\arguments{
\item{varname}{the name of the vector of numbers to be parsed}

\item{intxt}{the text data file read in using readLines}

\item{n}{the number of values to extract from the line}
}
\value{
a vector of n values, if the varname not present it returns NAgetvect
}
\description{
getvect parses a line of text and extracts 'n' pieces of
text as numbers. The specific line within the intxt, is identified by
the text at the start of the line found using varname
}
\examples{
txt1 <- "trawl,   1.4E-05,    2,    0.75,"  
txt2 <- "autoline,9.0E-06,    5,    1.25,"
tmp <- rbind(txt1,txt2)
print(getvect("autoline",tmp,3))
print(getvect("trawl",tmp,3))
}
