% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilfuns.R
\name{getvectors}
\alias{getvectors}
\title{getvectors exacts a chosen column from data.frames held in a list}
\usage{
getvectors(x, dfname, var)
}
\arguments{
\item{x}{a list of model outputs where each member of the list contains at
least one named data.frame or matrix with named columns.}

\item{dfname}{the name of the selected data.frame within each list member}

\item{var}{the name of the column from each data.frame to be combined.}
}
\value{
a matrix of the variable var
}
\description{
getvectors assumes one has a list of model outputs where each
model output contains matrices or data.frames of values. The aim is to
extract particular columns from particular data.frames from each list
member.
}
\examples{
models <- vector(mode="list",length=5)
names(models) <- 1:5
for (i in 1:5) {
  outmod <- matrix(rnorm(25,mean=5,sd=1),nrow=5,ncol=5,
                   dimnames=list(1:5,c("a","b","c","d","f")))
  models[[i]] <- list(outmod=outmod,index=i)
}
getvectors(models,dfname="outmod",var="d")
}
