% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmr_utils.R
\name{grsearch}
\alias{grsearch}
\title{grsearch a golden ratio search algorithm to estimate F in an ASPM}
\usage{
grsearch(f, interval, M, cyr, Byr, tol = 1e-09)
}
\arguments{
\item{f}{the function whose value needs to match cyr (catch in year)}

\item{interval}{the interval over which to search}

\item{M}{the instantaneous natural mortality rate}

\item{cyr}{the observed catch in a given year}

\item{Byr}{the exploitable biomass at the start of the year in which the
catches are to be taken}

\item{tol}{the tolerance of the match between the predicted and observed
catch}
}
\value{
an instantaneous fishing mortality rate that should generate the
observed catches from a given exploitable biomass in a year
}
\description{
grsearch implements a one-dimensional golden-ratio search for
the instantaneous F that gives rise to a required catch level in a
given year. It uses an input function and interval to search and a value
to match. The algorithm itself is described in many places on the www.
}
\examples{
matchC <- function(Fy,M,cyr,Byr) {
  return(abs((Byr * (1 - exp(-(M + Fy))) * Fy/(M + Fy)) - cyr))
}
grsearch(f=matchC,interval=c(0.2,0.45),M=0.036,cyr=5117.988,Byr=15760.748,
         tol = 1e-09)  # should be 0.4007973
}
