% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{IAconst}
\alias{IAconst}
\title{IAconst the first IA model example const object}
\format{
A list of five objects, biology, glb, fish, outrmse, and agecomp
}
\usage{
IAconst
}
\description{
IAconst is a list of 5 objects for a single tralw fleet fishery
that includes CPUE and age-composition data. The objects are 1. biology,
contains growth, weight-at-age, and maturity parameters. 2. glb the
globals object, 3. fish, the fishery data year, catch, CPUE, 4. outrmse,
an analysis of the expected weighting to apply to the CPUE data, and
5. agecomp, a matrix of age x years, of the counts by age in samples of
the catches.
}
\examples{
 data(IAconst)
 str(IAconst)
 print(IAconst$agecomp$twl[,1:8])
}
\keyword{datasets}
