% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{IArobust}
\alias{IArobust}
\title{IArobust are robustness outputs for 1 Fleet 1 region fishery}
\format{
A data.frame of 15 columns and 200 rows
}
\usage{
IArobust
}
\description{
IArobust a data.frame containing the results of running
a robustness test on the one fleet one region/stock fishery example.
The first Integrated Assessment that used catches, cpue,
and age-composition data, estimating 45 recdev parameters R0, q, sel50,
deltaS, and 41 recruitment deviates. This used the IAdynF function to
describe the stock dynamics using instantaneous fishing mortality rates.
It constitutes a 200 x 15 data.frame with columns iR0, iq, iselA50,
ideltaS, -veLL, eR0, eq, eselA50, edeltaS, R0, q, selA50, deltaS, dist,
and Iters. Where dist is the Euclidean distance between the initial
parameters and the original optimum. Each nominal parameter can be
changed by up to 8 percent. The results are sorted from
the smallest to the largest -veLL.
}
\examples{
 data(IArobust)
 str(IArobust)
 countLL <- table((round(IArobust[,"-veLL"],6))); length(countLL)
 countLL
 range(IArobust[,"-veLL"]) # should be 18.28771 18.75626
}
\keyword{datasets}
