% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.R
\name{initialdynamics}
\alias{initialdynamics}
\title{initialdynamics plots dynamics from initial parameter guesses}
\usage{
initialdynamics(
  x,
  year = "year",
  cpue = "cpue",
  predCE = "predCE",
  width = 9,
  height = 6,
  result = NULL,
  legcex = 1.25,
  legloc = "topright",
  ...
)
}
\arguments{
\item{x}{the input data, at least a matrix of 'year' and 'cpue'}

\item{year}{character name of the year variable, default = 'year'}

\item{cpue}{character name of the CPUE or index variable, default = 'cpue'}

\item{predCE}{character name of the predicted CPUE,  default = 'predCE'}

\item{width}{default = 9, the width of the plot}

\item{height}{default = 6, the height of the plot}

\item{result}{the BO, total likelihood and cpueLL as a column matrix.
default = NULL, which means nothing added to plot}

\item{legcex}{default = 1.25 the font size for the legend if used}

\item{legloc}{thge location of the legend if used. default = 'topright'}

\item{...}{other potential inputs, plotting parameters, etc.}
}
\value{
nothing but it does produce a plot
}
\description{
initialdynamics generates a plot of the predicted stock
cpue given a set off initial parameters. This can be used to find,
using trial and error, a set of scaling parameters (R0 and q, and
sometimes selectivity), that keep the predicted stock cpue off the
zero line and that intersect with the observed CPUE. Only plots
one fleet at a time.
}
\examples{
# think of something 
}
