% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agestruct.R
\name{initiateglobals}
\alias{initiateglobals}
\title{initiateglobals - calculates fishery constants as vectors}
\usage{
initiateglobals(consts, selabove = 0)
}
\arguments{
\item{consts}{the output from 'getconstants'}

\item{selabove}{are there ages or sizes that are never selected. Identify
the upper index in the outsel vector here. eg. if age 0 is never selected
then set zerocomps = 1, if ages 0-3 then zerocomps = 4}
}
\value{
a list containing LaA, WaA, Sel, MatA, growtran, and L50, and more
}
\description{
initiateglobals - calculates fishery constants as vectors,
including length-at-age, weight-at-age, selectivity-a-age, length-at
50\% selection, maturity-at-age, and a growth transition matrix. It now
includes sex, fleets, and regions as partitions of the model.
}
\examples{
data(const)
globals <- initiateglobals(const)
str(globals,max.level=1)
}
