% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{knifelogistic}
\alias{knifelogistic}
\title{knifelogistic a Logistic selectivity function with knifeedge option}
\usage{
knifelogistic(L50, delta, lens, knifeedge = 0, maxLML = 0)
}
\arguments{
\item{L50}{is the length at 50 percent selection}

\item{delta}{is the difference between the 95 percent selection and
the 50 percent selection}

\item{lens}{a vector of lengths for which the logistic maturity value
will be calculated}

\item{knifeedge}{defaults to 0. If knifeedge is set to a particular
length then the selectivity less than the value of knifeedge is set
to zero.}

\item{maxLML}{default = 0. The parameter is to allow for a slot selectivity}
}
\value{
A vector of length(lens) containing the predicted selectivity at
length values
}
\description{
knifelogistic a logistic selectivity function with the option
of including a knifeedge to allow for LML. This uses the
logistic function 1/(1+exp(-log(19.0)*(lens-inL50)/(delta))),
where delta = inL95 - inL50. This explicitly defines the SM50
but uses delta (which is SM95-SM50) as the second. This ensures
that when adding variation to parameters, to vary between
populations, when SM95 and SM50 are close together it is not
possible for SM50 to become larger than SM95. Be careful using the
knifeedge option. Strictly knifeedge selectivity would entail the
selectivity values being zero up to the knife-edge and then being 1.0.
This is not what happens here. Instead the knifeedge option literally
sets all values to zero at and below the value of knifeedge but leaves
any curve above that value as it is. Hence this is not strict knife-edge
selectivity. However, it does provide a selectivity curve that reflects
the selectivity of a fishery that uses a Legal Minimum Length or Size.
}
\examples{
\dontrun{
inL50 <- 100.0
delta <- 8.0
lens <- seq(2,210,2)
select <- knifelogistic(L50=inL50,delta,lens)
select <- knifelogistic(L50=inL50,delta,lens,knifeedge=105)
select <- knifelogistic(L50=inL50,delta,lens,maxLML=185)
}
}
