% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilfuns.R
\name{listtotext}
\alias{listtotext}
\title{listtotext is used in makehtml to make a text summary of a list}
\usage{
listtotext(x, sampn = 4)
}
\arguments{
\item{x}{the list to be described}

\item{sampn}{if an object in the list is numeric or a matrix this number
determines how large a sample of the contents are given. If nobs is
larger than the number of columns then the sample size becomes the
number of columns}
}
\value{
a vector or character vectors
}
\description{
listtotext is a utility used to summarize the contents of a
list of objects and their properties for inclusion when using makehtml
to display results. If, for example, one wants to report the structure
of a list using the str1 function this converts the list object into a
vector of character vectors that describes the contents. Inner list
objects are names while matrices have their dimension and a sample of
their contents
}
\examples{
eg <- list(biology=c(Linf=50,K=0.25,t0=-0.5,M=0.2),glb=list(a=2,b=3,c=4))
}
