% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmr_utils.R
\name{logistic}
\alias{logistic}
\title{logistic standard selectivity function}
\usage{
logistic(inl50, inl95, depend)
}
\arguments{
\item{inl50}{is the length/age at 50 percent selection/maturity/whatever}

\item{inl95}{is the length/age at 5 percent selection/maturity/whatever}

\item{depend}{a vector of lengths/ages for which the logistic value will be
calculated.}
}
\value{
A vector of length/age(depend) containing predicted logistic values
}
\description{
logistic calculates a Logistic curve that can be used as a
selectivity function, or maturity curve, of wherever a logistic is
required. This version uses the logistic function
1/(1+exp(-log(19.0)*(lens-inL50)/(inL95-inL50))),
which explicitly defines the SM50 and uses SM95 as the second parameter.
}
\examples{
in50 <- 3.398
in95 <- 4.386
ages <- seq(2,10,1)
select <- logistic(inl50=in50,inl95=in95,depend=ages)
round(cbind(ages,select),5)
}
