% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{makeabstock}
\alias{makeabstock}
\title{makeabstock defines an abalone stock's properties prior to assessment}
\usage{
makeabstock(fish, const, glb, pin, listtoenv = FALSE)
}
\arguments{
\item{fish}{the fishery dependent data including year, LML, catch, and cpue}

\item{const}{parameters for weight-at, maturity-at, and emergence-at-size,
and the von Bertalanffy parameters Linf, K, and CV.}

\item{glb}{the global constants object}

\item{pin}{the vector of initial parameters and whether they are fitted or
fixed}

\item{listtoenv}{default=FALSE, should list2env be used to put all list
objects into the .GlobalEnv.}
}
\value{
a list of size objects, biol, B0, dyn, Nt, NEbeg, and fishery
}
\description{
makeabstock characterizes an abalone stock and fishery prior to
its use in a stock assessment. The output list contains a 'biol' list
containing the weight, maturity and emergence from crypsis for each
size/length or size-class. It also contains the growth transition matrix.
It also contain the unfished biomass B0, then the initial fishery
dynamics 'dyn', containing catch, cpue, mature and exploitable biomass,
etc. Then 'Nt' the matrix that will contain the numbers-at-time,
containing the initial numbers-at-size in the first column. The 'NEbeg',
the start of year numbers-at-size of exploitable abalone. The final
'fishery' object contains each year's selectivity (to allow for LML
changes through time), the unique LMLs that occur, the years in the data,
the number of years, the survivorship from natural mortality and that
from half the natural mortality.
}
\examples{
data(absaudata)
abstk <- makeabstock(fish=absaudata$fish,const=absaudata$const,
                     glb=absaudata$glb,pin=absaudata$pindat,listtoenv=FALSE)
str(abstk)
}
