% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{makebiol}
\alias{makebiol}
\title{makebiol defines an abalone stock's biological properties}
\usage{
makebiol(fish, const, glb, listtoenv = TRUE)
}
\arguments{
\item{fish}{the fishery dependent data including year, LML, catch, and cpue}

\item{const}{parameters for weight-at, maturity-at, and emergence-at-size,
and the von Bertalanffy parameters Linf, K, and CV.}

\item{glb}{the global constants object}

\item{listtoenv}{default=TRUE, should list2env be used to put all list
objects into the .GlobalEnv.}
}
\value{
a list of the matrix of biological properties and G the growth
transition matrix
}
\description{
makebiol characterizes an abalone stock prior to
its use in a stock assessment. The output list contains a 'biol' list
containing the weight, maturity, emergencefrom crypsis, and the combined
matwt for each size/length or size-class. It also contains the growth
transition matrix. By default the result is returned invisibly but is
also returned to the gloabl environment for immediate use
}
\examples{
data(absaudata)
biolout <- makebiol(fish=absaudata$fish,const=absaudata$const,
                     glb=absaudata$glb,listtoenv=FALSE)
str(biolout)
# fish=absaudata$fish;const=absaudata$const;glb=absaudata$glb
}
