% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agestruct.R
\name{makedataset}
\alias{makedataset}
\title{makedataset generates a csv file from siumulated data}
\usage{
makedataset(
  glb,
  out,
  randseed = NULL,
  catchSD = 1e-08,
  cpueSD = 1e-08,
  getA = FALSE,
  sampnA = 500,
  sampsdA = 50,
  maxnA = 1000,
  redvarA = 50,
  getL = FALSE,
  sampnL = 800,
  sampsdL = 50,
  maxnL = 2000,
  redvarL = 50
)
}
\arguments{
\item{glb}{the output object from initiateglobals}

\item{out}{the output object from simdynF}

\item{randseed}{default=NULL, otherwise this value is used to set.seed to
obtain repeatable outputs}

\item{catchSD}{standard deviation of normal variation imposed on reported
catches. default = 1e-08, effectively no variation}

\item{cpueSD}{standard deviation of normal variation imposed on reported
CPUE. default = 1e-08, effectively no variation}

\item{getA}{default=FALSE, sampling includes age-composition of catches. If
getA==TRUE, then next four params are active}

\item{sampnA}{expected booststrap age sample = 500}

\item{sampsdA}{sample size sd used by rowfreqboot default=50}

\item{maxnA}{the maximum sample size possible (> n), default = 1000}

\item{redvarA}{sd for how much each column of catch numbers are reduced
prior to bootstrapping, deault =50}

\item{getL}{default=FALSE, sampling includes age-composition of catches. If
getA==TRUE, then next four params are active}

\item{sampnL}{expected booststrap age sample = 800}

\item{sampsdL}{sample size sd used by rowfreqboot default=50}

\item{maxnL}{the maximum sample size possible (> n), default = 2000}

\item{redvarL}{sd for how much each column of catch numbers are reduced
prior to bootstrapping, deault =50}
}
\value{
a list of histC, agecomp, and sizecomp, some of which may be NULL
}
\description{
makedataset generates a csv file from simulated data produced
by getconstants, initiateglobals, and simdynF.
}
\examples{
\dontrun{
 const2 <- getconstants(pathtopath(datraw,"F2-A1-S1.csv"))
 glb <- initiateglobals(consts=const2,selabove=1)
 stock <- makestock(glb=glb)
 out <- simdynF(glb=glb,stk=stock,reps=5,full=TRUE)
 catchSD=1e-08
 cpueSD=1e-08
 sigR=1e-08
 outdata <- makedataset(glb,out,catchSD=1e-08,cpueSD=1e-08,sigR=1e-08)
 # now open 'F21S.csv
}
}
\seealso{
{
\link{getconstants}, \link{initiateglobals}. \link{simdynF}
}
}
