% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agestruct.R
\name{makerecdev}
\alias{makerecdev}
\title{makerecdev generates a single recruitment deviate for a given sigmaR}
\usage{
makerecdev(sigmaR, n = 1)
}
\arguments{
\item{sigmaR}{simulated recruitment variability}

\item{n}{number of recdevs to generate}
}
\value{
a single bias-corrected recruitment deviate
}
\description{
makerecdev is used to include recdevs into a stock recruitment
relationship, as in  rec <- bhsim * makerecdev(sigmaR)
}
\examples{
makerecdev(sigmaR=0.5)
}
