% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agestruct.R
\name{makesimstock}
\alias{makesimstock}
\title{makesimstock sets up the containers for the stock dynamics}
\usage{
makesimstock(glb)
}
\arguments{
\item{glb}{the constants obtained from constants.csv including those
relating to the biology, the fishery, and the structural details.}
}
\value{
a list containg the more important placeholders for the stock and
fishery dynamics.
}
\description{
makestock sets up the containers for the stock dynamics. The
objective is to have an object that can be passed to substantive
functions so the dynamics can be followed easily without using the
global environment and remembering all the particular variables.
}
\examples{
library(fmr)
data(const)
globals <- initiateglobals(const)
stock <- makesimstock(globals)
str(stock)
}
