% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{makestock}
\alias{makestock}
\title{makestock set up the size-based model stock}
\usage{
makestock(fish, const, glb, pin)
}
\arguments{
\item{fish}{the fishery dependent data year, LML, catches cpue}

\item{const}{the biological constants relating to maturity, growth, and
weight-at-length}

\item{glb}{the globals object}

\item{pin}{the initial starting values of the estimable parameters and
whether they are estimated or not.}
}
\value{
a list of the biological properties, B0, matB, expB, Nt, NEbeg,
depl, and fishery details.
}
\description{
makestock initiates a size-based model of a stock by using the
biological and fishery parameters to estimate the required maturity,
selectivity, weight-at-length, unfished numbers-at-size, unfished mature
and exploitable biomass.
}
\examples{
print("wait on data sets")
}
