% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{maplike}
\alias{maplike}
\title{maplike the likelihood surface of LnR0 and Lnq for ASPM}
\format{
A 2601 row x 12 column data.frame
}
\usage{
maplike
}
\description{
A 2601 row data.frame containing the initial parameter vector,
the -veLL, the optimum solution, the back transformed optimum, the
minimum gradient, and a counter. This can be used to plot the lieklihood
surface in a crude manner it ranges from LnR0 = seq(5.9,8.4,0.05) and
Lnq = seq(-8.7,-6.2,0.05), which is 51 x 51.
}
\examples{
data(maplike)
pickbad <- which(maplike[,"-veLL"] > 3)
roundlike <- round(maplike[,"-veLL"],8)
outn <- table(roundlike); print(outn)
pickopt <- which(roundlike == as.numeric(names(outn)[1]))
pickworst <- which(roundlike == as.numeric(names(outn)[length(outn)])) 
plot(maplike[,3],maplike[,1],type="p",cex=1,col=0,xlab="Ln(q)",
     ylab="Ln(R0)",panel.first=grid())
points(maplike[pickbad,3],maplike[pickbad,1],cex=2,col=2)
points(maplike[pickopt,3],maplike[pickopt,1],cex=3,col=5)
points(maplike[pickworst,3],maplike[pickworst,1],cex=3,col=3)
}
\keyword{datasets}
