% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{negLL}
\alias{negLL}
\title{negLL estimates the negative log-likelihood for an abalone stock}
\usage{
negLL(
  pars,
  funk,
  initpar = pars,
  glb,
  fish,
  biol,
  G,
  const,
  notfixed = c(1:length(pars)),
  omega,
  finalcomp,
  full = FALSE,
  initH = 0
)
}
\arguments{
\item{pars}{the log-transformed model parameters}

\item{funk}{the dynamics function, for abalone = abdynaicsH}

\item{initpar}{a copy of the initial parameters}

\item{glb}{the globals object}

\item{fish}{the fishery data including year, LML, catch and cpue}

\item{biol}{the biol matrix from makebiol}

\item{G}{the growth transition matrix from makebiol}

\item{const}{the biological constants relating to weight-, maturity-,
emergence, and growth-at-size}

\item{notfixed}{a vector of the indices of the parameters to be estimated}

\item{omega}{the weights to be applied to each data sets log-likelihood}

\item{finalcomp}{a matrix containing only those columns of the sizecomp data
that are to be included in the analysis}

\item{full}{should the full details be generated or only the final -veLL}

\item{initH}{an initial harvest rate used to generate an initial depletion}
}
\value{
Either just the log-likelihood or a vector of its components.
}
\description{
negLL given an abalone stock along with all the related
requirements, negLL estimates the -ve Log-Likelihood that combines the
-veLL from the cpue and from the size composition data.
It can output either only the -veLL (used when fitting the model to data)
or a listing of the components of the likelihood and other constants
}
\examples{
\dontrun{
 data(absaudata)
 list2env(absaudata,envir=.GlobalEnv)
 out <- abdynamicsH(pin=pindat[,1],fish=fish,biol=biol,G=G,const=const,
                    glb=glb)
 round(out$dyn,3)
 picksc=which(colSums(sizecomp) > 100)
 finalcomp <- sizecomp[,picksc]
 notfixed=which(pindat[,2] > 0)  
 negLL(pars=pindat[,1],funk=abdynamicsH,initpar=pars,glb=glb,fish=fish,
       biol=biol,G=G,const=const,notfixed=notfixed,
       omega=c(1,1e-03),finalcomp=finalcomp,full=FALSE,initH=0)
}
#  pars=pin;initpar=allpin;method="BFGS";
#  funk=abdynamicsH;glb=glb;fish=fish;biol=biol;G=G;const=const
#  omega=omega;finalcomp=sizecomp;notfixed=notfix;full=FALSE;initH=0
}
