% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ocaa}
\alias{ocaa}
\title{ocaa the observed catch-at-age for plaice.}
\format{
A 9 x 9 data.frame
}
\usage{
ocaa
}
\description{
Table 13.1 from Beverton & Holt, 1957, p451 containing the relative
catch-at-age for plaice from the north sea. This is a data.frame of nine
years (29/30 - 37/38) by 9 ages (2 - 10).
}
\examples{
data(ocaa)
ocaa
}
\references{
Beverton, R.J.H. and S.J. Holt (1957) \emph{On the dynamics of
exploited fish populations.} U.K. Ministry of Agriculture and Fisheries,
Fisheries Investigations (Series 2), \emph{19}: 1-533.
}
\keyword{datasets}
