% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{owaa}
\alias{owaa}
\title{owaa the observed weight-at-age for plaice.}
\format{
A 9 x 9 data.frame
}
\usage{
owaa
}
\description{
Table 16.2 from Beverton & Holt, 1957 containing the relative
weight-at-age for plaice from the north sea. This is a data.frame of nine
years (29/30 - 37/38) by 9 ages (2 - 10). Older ages are omitted.
}
\examples{
data(owaa)
owaa
}
\references{
Beverton, R.J.H. and S.J. Holt (1957) \emph{On the dynamics of
exploited fish populations.} U.K. Ministry of Agriculture and Fisheries,
Fisheries Investigations (Series 2), \emph{19}: 1-533.
}
\keyword{datasets}
