% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{param}
\alias{param}
\title{param preliminary parameter estimates for a SCAA model of plaice}
\format{
A data.frame of two columns
\describe{
\item{var}{ the name of each parameter }
\item{value}{ the natural logarithm transformed value of each parameter }
}
}
\usage{
param
}
\description{
Preliminary parameters used when fitting a statistical catch-at-
age model to the plaice data from Bevertone & Holt, 1957. Much of this
data has also been included in the age-structured model described in
Haddon, 2011.
}
\examples{
 data(plaice)
 str(plaice)
 print(plaice$fish)
 print(plaice$agedata)
}
\references{
Beverton, R.J.H. and S.J. Holt (1957) \emph{On the dynamics of
exploited fish populations.} U.K. Ministry of Agriculture and Fisheries,
Fisheries Investigations (Series 2), \emph{19}: 1-533.
Haddon, M (2011) \emph{Modelling and Quantitative Methods in Fisheries}
2ns edition. Chapman & Hall, CRC Press. Boca Raton. 449p.
}
\keyword{datasets}
