% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pindat}
\alias{pindat}
\title{pindat preliminary parameter estimates for a SBM model for abalone}
\format{
A data.frame of two columns
\describe{
\item{var}{ the name of each parameter }
\item{value}{ the natural logarithm transformed value of each parameter }
}
}
\usage{
pindat
}
\description{
Preliminary parameters used when fitting a statistical catch-at-
size model to the abalone data from McAllister & Mundy (2024). This
data has also been included in the size-based MSE, 'aMSE' described at
https://haddonm.github.io/aMSEGuide
}
\examples{
 data(plaice)
 str(plaice)
 print(plaice$fish)
 print(plaice$agedata)
}
\references{
McAllister, J. and C. Mundy (2024) \emph{Tasmanian Abalone
Assessment}, Institute for Marine & Antarctic Studies, University of
Tamsania p 1-144.
}
\keyword{datasets}
