% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fmr.R
\name{plot_fmr}
\alias{plot_fmr}
\title{Do all fmr plots}
\usage{
plot_fmr(M, plot_res = FALSE, target_dir = "", fmr_options = .fmr_options)
}
\arguments{
\item{M}{a list object created by read_admb function}

\item{plot_res}{plot residuals only (default = FALSE)}

\item{fmr_options}{the default plotting options}
}
\value{
All fmr plots.
}
\description{
Do all fmr plots
}
\author{
DN Webber
}
