% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.R
\name{plotASPM}
\alias{plotASPM}
\title{plotASPM plots catch, CPUE, Spawning Biomass and Harvest Rate}
\usage{
plotASPM(
  infish,
  CI = NA,
  target = 0.48,
  usef = 7,
  rundir = "",
  console = TRUE,
  scenario = "",
  maxy = 0
)
}
\arguments{
\item{infish}{an object generated by the dynamics function}

\item{CI}{defaults to NA, if confidence intervals around the cpue have been
obtained using getLNCI, then the resulting matrix will generate 95pc CIs}

\item{target}{target depletion level. Defaults to 0.48}

\item{usef}{defines the font to use usef(ont),default = 7 bold times}

\item{rundir}{default '', otherwise give a full path if saving a file}

\item{console}{default=TRUE, which implies the plot will go to console. If
set TRUE it will save pathtopath(rundir,'scenario.png')}

\item{scenario}{default = '', but if a file is to be saved then this should\
be given a name for the file, if left as '', then it will be called
ASPM.png}

\item{maxy}{default = 0, which means use the data to set ymax. If maxy > 0
the ymax will be set to that value. Used to allow details to be seen}
}
\value{
Nothing, but it does plot six graphs in a single plot.
}
\description{
plotASPM after running fitASPM the optimum parameters can be
put through the dynamics function to generate a dataframe containing
the optimum dynamics. These can be plotted using plotASPM, which plots
out the catches, the Spawning Biomass, the relative CPUE and its fit to
the observed CPUE, and the harvest rate. This routine is still under
development to include more options.
}
\examples{
require(fmr)
data(westroughy)
fish <- westroughy$fish
glb <- westroughy$glb
props <- westroughy$props
pars <- c(7.1,-1,-7.7)  # logR0 logceSE logavq
bestaspm <- fitASPM(pars,minfun=dynamicsH,infish=fish,
                   inglb=glb,inprops=props)
out <- dynamicsH(bestaspm$estimate,fish,glb,props,full=TRUE)
plotASPM(out$fishery, console=TRUE)
ceCI <- getLNCI(out$fishery[,"predCE"],exp(bestaspm$estimate[2]))
plotASPM(out$fishery,CI=ceCI, console=TRUE)
# infish=fishery; CI=ceCI;defineplot=TRUE; target=0.48; usef=7;rundir=""
# console=TRUE; scenario=""; maxy=10
}
