% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makesbmresults.R
\name{plotbiology}
\alias{plotbiology}
\title{plotbiology generates a pot of the maturity, emergence, WtL, and MatWt}
\usage{
plotbiology(rundir, biol, console = TRUE)
}
\arguments{
\item{rundir}{the directory for all ctrl, data, and output files.}

\item{biol}{the matrix of biological properties (except growth).}

\item{console}{should the plot go to the console or be saved? Default=TRUE}
}
\value{
nothing but it does generate a plot to the console or to rundir
}
\description{
plotbiology only generates a 2 x 2 plot of maturity-at-length,
emergence-at-length, weight-at-length, and a combination of maturity
and weight-at-length (to simplify a few repeated calculations). If a
filen is given then the plotfile is automatically stored in rundir
and prepared for inclusion in the results webpage.
}
\examples{
print("wait on data sets")
#   plotbiology(rundir=resdir,biol=biol,console=TRUE)
}
\seealso{
{
\link{makebiol}, \link{makehtml}
}
}
