% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.R
\name{plotcompfit}
\alias{plotcompfit}
\title{plotcompfit plots observed vs predicted composition proportions}
\usage{
plotcompfit(
  obscomp,
  predcomp,
  analysis = "Composition_Fit",
  ylabel = "opt49_ages",
  console = TRUE,
  outdir = "",
  predcol = "red",
  cex = 1,
  topcex = 0.9,
  maxcat = 0
)
}
\arguments{
\item{obscomp}{the observed composition data from the fishery as a matrix
with ages or sizes as rownames and years as column names}

\item{predcomp}{the predicted composition data from the model as a matrix
with ages or sizes as rownames and years as column names. Only the years that match the observed years will be plotted.}

\item{analysis}{a character name for the comparison being made,
default='composition_fit'. Whatever is used should have no spaces as it is
also used in teh filename, if used}

\item{ylabel}{a character label for the y-axis, this is added to the
contents of the analysis argument, default='opt49_ages'}

\item{console}{should the plot got to the console or file, default = TRUE}

\item{outdir}{the directory into which to send the file is console = FALSE,
default =''}

\item{predcol}{the plotted colour of the predicted line, default = 'red'}

\item{cex}{the generic size of the font used in the plots}

\item{topcex}{the font size used for the top labels = year and sample size}

\item{maxcat}{the maximum age or size to be plotted. Default = 0, which
implies all catcegories are plotted. This can be useful when each plot
has a long tail of very few observations.}
}
\value{
It generates a plot and, invisibly, a list of the filename and
caption in case it is saved as a file, as well as the final obscomp
and predcomp, if differences are to be plotted.
}
\description{
plotcompfit generates a plot of the predicted proportional
composition in the catches versus the observed proportional composition
data to illustrate the fit of the model to the observed data. The xlabel
of all plots is left blank, but the range of each x-axis is printed at
the bottom of all.
}
\examples{
\dontrun{
  predcomp <- opt49$catchN[,2:46]
  obscomp <- const$agecomp$twl
  plotcompfit(obscomp,predcomp,analysis="Model Fit",ylabel="opt49",
              console=TRUE,outdir="",predcol="red",xlabel="",cex=1.0,
              topcex=0.9,maxcat=15)
}
}
