% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.R
\name{plotjitter}
\alias{plotjitter}
\title{plotjitter generates plots of single fishery variables from a jitter}
\usage{
plotjitter(
  mods,
  pickvar,
  label,
  defplot = FALSE,
  addvect = NULL,
  doplot = TRUE
)
}
\arguments{
\item{mods}{the list of model outputs from each N trial}

\item{pickvar}{the character name of the column to be plotted}

\item{label}{the y-axis label of the plot, if empty = pickvar}

\item{defplot}{should a single plot be generated of will it be part of a
multipanel plot. If TRUE a single plot is produced.}

\item{addvect}{a vector of values to be plotted on top of the N replicates.
default = NULL}

\item{doplot}{should a plot be produced, default = TRUE. If FALSE only the
extracted pickvar matrix is produced.}
}
\value{
nothing but it does produce a plot
}
\description{
plotjitter enables each variable from the fishery data output
from a jitter analysis to be plotted to illustrate any degree of
variation.
}
\examples{
\dontrun{
  mods <- models29init
  N <- length(mods)
  plotprep(width=10,height=8)
  parset(plots=c(2,2),cex=1.0)
  plotjitter(mods,pickvar="spawnB",label="Spawning Biomass t",defplot=FALSE)
  plotjitter(mods,pickvar="deplete",label="Spawning Biomass Depletion",
             defplot=FALSE)
  plotjitter(mods,pickvar="twlPCE",label="Trawl CPUE",defplot=FALSE,
             addvect=fish$twlCE)
  plotjitter(mods,pickvar="recruit",label="Recruitment",defplot=FALSE)
}
}
