% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{plotSBMdynamics}
\alias{plotSBMdynamics}
\title{plotSBMdynamics plots the implied dynamics from the assessment}
\usage{
plotSBMdynamics(rundir, dyn, fish, pars, glb, console = TRUE)
}
\arguments{
\item{rundir}{the directory for all ctrl, data, and output files.}

\item{dyn}{the matrix of the dynamics out of the 'dynamics' function
that uses the optimum estimate of the parameters}

\item{fish}{the fishery data used in the assessment}

\item{pars}{the back-transformed parameters from the assessment}

\item{glb}{the globals object}

\item{console}{should the plot go to the console or be saved? Default=TRUE}
}
\value{
nothing but it does add a plot to the console or to rundir
}
\description{
plotSBMdynamics generates a 3 x 2 plot of the cpue imposed on
the predicted cpue, the exploitable biomass, the mature biomass,
the harvest rate, the recruitment levels, and the mature biomass
depletion levels.
}
\examples{
\dontrun{
    plotSBMdynamics(rundir,dyn,fish,exp(pindat[,1]),glb,console = TRUE)
}
#  rundir=rundir; console=TRUE; pars=allpin
# rundir=rundir;dyn=outdyn$dyn;fish=fish;pars=exp(out$optpar);glb=glb;
# fisindex=NULL; predfisindex=NULL; console = TRUE
}
