% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{preliminaryfit}
\alias{preliminaryfit}
\title{preliminaryfit obtains the fit to inital parameter values}
\usage{
preliminaryfit(
  pindat,
  glb,
  fish,
  biol,
  G,
  const,
  sizecomp,
  omega,
  full = TRUE,
  initH = 0
)
}
\arguments{
\item{pindat}{the matrix containing the vector of parameter values and the
vector of phases - either 1 or 0 indicating which will be estimated and
which fixed}

\item{glb}{the globals object}

\item{fish}{the fishery data including year, LML, catch and cpue}

\item{biol}{the biol matrix from makebiol}

\item{G}{the growth transition matrix from makebiol}

\item{const}{the biological constants relating to weight-, maturity-,
emergence, and growth-at-size}

\item{sizecomp}{the size composition data from commercial catches}

\item{omega}{the weights to be given to each data stream}

\item{full}{should full details be given or just the log-likelihood}

\item{initH}{an initial harvest rate used to generate an initial depletion}
}
\value{
Either just the log-likelihood or a vector of its components. It
also generates a plot of the initial dynamics
}
\description{
preliminaryfit runs the dynamics and pots them as well as
calculating the initial negative log-likelihood
}
\examples{
\dontrun{
 data(absaudata)
 list2env(absaudata,envir=.GlobalEnv)
 biolout <- makebiol(fish=fish,const=const,glb=glb)
 out <- abdynamicsH(pin=pindat[,1],fish=fish,biol=biol,G=G,const=const,
                    glb=glb)
 round(out$dyn,3)
 oldlogL <- preliminaryfit(pindat=pindat,glb=glb,fish=fish,biol=biol,G=G,
                           const=const,sizecomp=sizecomp,omega=c(1,1e-3),
                           full=FALSE,initH=0)
 oldlogL
}
}
