% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilfuns.R
\name{propdiff}
\alias{propdiff}
\title{propdiff gives the proportional difference of the range of a vector}
\usage{
propdiff(x, var, giverge = TRUE)
}
\arguments{
\item{x}{a list containing multiple named scalar values or a numeric vector}

\item{var}{the name of the scalar to be extracted from each member of the
list and converted into a vector of values. If the input is a vector,
this argument can be ''.}

\item{giverge}{default = TRUE, which means the range of values will be
returned as well as the propdiff. If FALSE, only the propdiff is
returned.}
}
\value{
either the range of values and the proportional difference between
the largest and smallest, or just the proportional difference of the
range.
}
\description{
propdiff takes in either a list of objects all of which
contain named scalars, or a numeric vector of numbers and calculates
the range of the values and then outputs the proportional
difference between the smallest and the largest. So the result is the
proportional difference between the largest and smallest divided by the
largest. This is useful when conducting a jitter analysis of fitting a
model, each of which uses a set of likelihood components.
}
\examples{
  x <- vector(mode="list",length=10)
  names(x) <- 1:10
  for (i in 1:10) x[[i]] <- c(value=rnorm(1,mean=5, sd=1),n=1)
  print(x)
  propdiff(x,var="value")
  x <- rnorm(10,mean=5, sd=1)
  propdiff(x,var="",giverge=FALSE)
}
