% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{readSBMctrl}
\alias{readSBMctrl}
\title{readSBMctrl reads the control file and generates working R object}
\usage{
readSBMctrl(rundir, infile = "controlLBM.csv", verbose = TRUE)
}
\arguments{
\item{rundir}{the full path to the working directory being used}

\item{infile}{the name of the control file found in rundir}

\item{verbose}{should progress be reported to the console, default=TRUE}
}
\value{
a list containing the ctrl, constants, and glb objects
}
\description{
readSBMctrl reads in the control file and outputs the ctrl,
constants, and glb objects
}
\examples{
\dontrun{
 rundir <- tempdir()
 SBMctrltemplate(rundir,infile="ctrlSBM.csv")
 setup <- readSBMctrl(rundir,infile="ctrlSBM.csv") 
}
}
\seealso{
{
\link{SBMctrltemplate}
}
}
