% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{readSBMdata}
\alias{readSBMdata}
\title{readSBMdata reads the fishery and sizecomp data for a SBM model}
\usage{
readSBMdata(rundir, infile)
}
\arguments{
\item{rundir}{the complete path to the directory containing the scenario
files}

\item{infile}{the name of the csv file containing the data}
}
\value{
a list containing a matrix of the fishery data, fish, and an array
of the size-composition data size-class x year x gender, sizecomp.
}
\description{
readSBMdata is used to read in the fishery data - year, catch,
cpue, and size composition data. The format for the data file can be
obtained by using the SBMdatatemplate function.
}
\examples{
print("wait on available datasets") #
}
