% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{readSBMpin}
\alias{readSBMpin}
\title{readLBMpin reads in an initial parameters file}
\usage{
readSBMpin(rundir, infile)
}
\arguments{
\item{rundir}{the full path to the working directory being used}

\item{infile}{the name of the pin file}
}
\value{
a vector of parameters in a fixed order
}
\description{
readSBMpin reads in the initial parameter values from the
pin file named in the control file and expected to be found in rundir.
Its format is very specific and the SBMpintemplate function should be
used to provide an example.
}
\examples{
print("wait on suitable internal data")
}
\seealso{
{
\link{SBMpintemplate}
}
}
