% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{resrtmb}
\alias{resrtmb}
\title{resrtmb contains robustness outputs for a 28P IA using rtmb}
\format{
A matrix of 14 columns and 200 rows
}
\usage{
resrtmb
}
\description{
resrtmb is a 200 x 14 matrix containing the results of running
a robustness test on the one fleet one region/stock fishery example
where 28 parameters are fitted to the data using the RTMB based function
IAdyntmbF. The parameters are  R0, q, selA50, deltaS, and 24 recruitment
deviates. The IAdyntmbF function describes the stock dynamics using
instantaneous fishing mortality rates. It constitutes a 200 x 14 matrix
with columns iR0, iq, iselA50, ideltaS, -veLL, eR0, eq, eselA50,
edeltaS, R0, q, selA50, deltaS, and Iters. All -veLL values are the same
but there is variation in all the other variables, though very small in
the estiamted parameters.
}
\examples{
 data(resrtmb)
 str(resrtmb)
 range(resrtmb[,"-veLL"]) # should be 348.7424 348.7424 ie no difference
 rge <- apply(resrtmb,2,range)
 rge1 <- t(rbind(rge,rge[2,] - rge[1,]))
 colnames(rge1) <- c("min","max","diff")
 print(round(rge1,5))
}
\keyword{datasets}
