% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{resrtmb29}
\alias{resrtmb29}
\title{resrtmb29 contains robustness outputs for a 29P IA using rtmb}
\format{
A matrix of 14 columns and 200 rows
}
\usage{
resrtmb29
}
\description{
resrtmb29 is a 200 x 14 matrix containing the results of running
a robustness test on the one fleet one region/stock fishery example
where 29 parameters are fitted to the data using the RTMB based function
IAdyntmbF. The parameters are  R0, q, selA50, deltaS, and 25 recruitment
deviates. The IAdyntmbF function describes the stock dynamics using
instantaneous fishing mortality rates. It constitutes a 200 x 14 matrix
with columns iR0, iq, iselA50, ideltaS, -veLL, eR0, eq, eselA50,
edeltaS, R0, q, selA50, deltaS, and Iters. All -veLL values are very
close in value with the maximum difference being 0.0002432351, there is
more variation in the other variables, though very small in
the estimated parameters.
}
\examples{
 data(resrtmb29)
 str(resrtmb29)
 minLL <- resrtmb29[1,"-veLL"]   # should be 173.5056
 maxLL <- resrtmb29[200,"-veLL"] # should be 173.5059
 { cat("min  = ",minLL,"\n")
   cat("max  = ",maxLL,"\n")
   cat("diff  = ",maxLL-minLL,"\n")
 }
}
\keyword{datasets}
