% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspm.r
\name{robustASPM}
\alias{robustASPM}
\title{robustASPM conducts a robustness test on the quality of fit of an ASPM}
\usage{
robustASPM(
  inpar,
  dynfun,
  fish,
  glb,
  props,
  N = 10,
  scaler = 15,
  Hrange = c(0.01, 0.45, 0.005),
  numyrs = 50,
  verbose = TRUE,
  steptol = 1e-05
)
}
\arguments{
\item{inpar}{the parameter set to begin the trials with, this must be a
named vector of the optimum parameter values = LogR0, logsigCE, logavq}

\item{dynfun}{the dynamics function giving rise to -veLL, usually dynF, or
dynamicsF, or dynamicsH}

\item{fish}{the fisheries data: at least year, catch, and cpue}

\item{glb}{the global variables containing the biological information}

\item{props}{the properties calculated from the globals}

\item{N}{the number of random trials to run; defaults to 10, which is too few}

\item{scaler}{the divisor that sets the degree of normal random variation to
add to the parameter values; default = 15 the smaller the value the more
variable the outcome}

\item{Hrange}{a vector of three numbers denoting the range of harvest rates
to use when characterizing the productivity implied by each fitted
parameter set. defaults to c(0.01, 0.45, 0.005)}

\item{numyrs}{the number of years used to drive each harvest rate to an
equilibrium population structure}

\item{verbose}{print progress count to the screen? default = TRUE}
}
\value{
a list of results from each run, the range of values across runs, and
the median values.
}
\description{
robustASPM conducts a robustness test on the quality of fit of
an ASPM. This is done by using the original optimal model parameters or
the original guessed parameter values, add random variation to each of
them, and re-fit the model. This process needs to be repeated multiple
times. This should enable an analysis of the stability of the modelling
outcomes. If the optimum parameters are used then add more variation, if
initial guesses are used you may need to select different starting points
so that the random variation covers the parameter space reasonably well.
}
\examples{
print("wait on data sets")
}
