% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{robustness}
\alias{robustness}
\title{robustness are 'robustASPM' outputs western Tasmanian Orange Roughy}
\format{
A data.frame of 17 columns and 200 rows
}
\usage{
robustness
}
\description{
robustresults a data.frame containing the results of running
the robustASPM function using the westroughy data sets and using the
dynamicsH function that has Annual Harvest rates. It
constitutes a 200 x 17 data.frame with columns iLnR0, isigmaCE, iavq,
-veLL, LnR0, LsigCE, Lavq, gradR0, gradsigCE, gradq, R0, sigCE, avq,
MSY, B0, pardist, and Iters.
See the description in the Introduction to Age-Structured Models
chapter in More Fisheries Modelling using R.
}
\examples{
 data(robustness)
 str(robustness)
 countLL <- table((round(robustness[,"-veLL"],6)))
 countLL
 range(robustness[,"-veLL"]) # should be 6.248774 7.039229
}
\keyword{datasets}
