% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{robustnessF}
\alias{robustnessF}
\title{robustnessF are 'robustASPM' outputs for roughy using instantaneous F}
\format{
A data.frame of 17 columns and 200 rows
}
\usage{
robustnessF
}
\description{
robustnessF is a data.frame containing the results of running
the robustASPM function using the westroughy data sets and using the
dynamicsF function that has instantaneous fishing mortality rates. It
constitutes a 200 x 17 data.frame with columns iLnR0, isigmaCE, iavq,
-veLL, LnR0, LsigCE, Lavq, gradR0, gradsigCE, gradq, R0, sigCE, avq,
MSY, B0, pardist, and Iters.
See the description in the Introduction to Age-Structured Models
chapter in More Fisheries Modelling using R. Though robustnessF is
generated in the Advanced ASPM chapter.
}
\examples{
 data(robustnessF)
 str(robustnessF)
 countLL <- table((round(robustnessF[,"-veLL"],6)))
 countLL
}
\keyword{datasets}
