% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{robustresults}
\alias{robustresults}
\title{robustresults are 'robustASPM' outputs western Tasmanian Orange Roughy}
\format{
A data.frame of 14 columns and 200 rows
\describe{
\item{fish}{ a data.frame containing year, catch, cpue, SE of the cpue }
\item{glb}{biological parameters relating to growth, selectivity,
weight-at-age, steepness, and resilience and spsname }
\item{props}{ contains six variables ages, laa, waa, maa, sela, and feca,
which are all relative to age.}
\item{agedata}{ a list of 5 objects, yrage - the years in which age data are
available, ages - the observed ages, agemax - the maximum age, nage -
the number of observed ages, and naa - the numbers-at-age by year}
\item{lendata}{ a list of 5 objects akin to the agedata object but for
length data.}
}
}
\usage{
robustresults
}
\description{
robustresults a data.frame containing the results of running
the robustASPM function using the westroughy data sets and using the
dynamicsF function that has instantaneous fishing mortality rates. It
constitutes a 200 x 14 data.frame with columns iLnR0, isigmaCE, iavq,
-veLL, LnR0, LsigCE, Lavq, R0, sigCE, avq, MSY, B0, pardist, and Iters.
See the description in teh Introduction to Age-Structured Models
chapter in More Fisheries Modelling using R.
}
\examples{
\dontrun{
 data(robustresults)
 str(robustresults)
}
}
\keyword{datasets}
