% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmr_utils.R
\name{rowfreqboot}
\alias{rowfreqboot}
\title{rowfreqboot generates a bootstrap sample of a matrix of counts}
\usage{
rowfreqboot(x, n)
}
\arguments{
\item{x}{a matrix or data.frame of sample counts of a set of integer
categories, such as ages or sizes, that are used as column headings.}

\item{n}{the bootstrap sample size for each row}
}
\value{
a matrix the same dimensions as x containing a bootstrap sample of x
}
\description{
rowfreqboot in fisheries biology it is common to have a matrix
of counts or ages or sizes or some other integer numeric property. It
would also, often be useful to be able to generate a bootstrap sample of
such data, either for simulation or to obtain bootstrap estimates of
variability. rowfreqboot expects the integer numeric values to be used
as the column headings (ages, sizes) and the rows to the separate sample
counts of the different column categories. It will then generate a matrix
of the same dimension as that input that contains an empirical bootstrap
sample of size n for each row
}
\examples{
dat <- c(400,1000,600,200,300,50,10,20,5,2,1000,236,556,304,91,125,19,4,7,2,
          700,595,135,297,151,43,55,8,2,3)
inmat <- matrix(dat, nrow=3,ncol=10,byrow=TRUE,dimnames=list(c(1:3),c(4:13)))
rowfreqboot(x=inmat,n=200)
rowfreqboot(x=inmat,n=200)
}
