% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agestruct.R
\name{samplefishery}
\alias{samplefishery}
\title{samplefishery generates catcges and CPUE with-without errors}
\usage{
samplefishery(out, glb, errors = c(catchSD = 1e-08, cpueSD = 1e-08))
}
\arguments{
\item{out}{the output of simdynF, which generates the dynamics of the
simulated fishery}

\item{glb}{the globals object in teh simulation, from initiateglobals}

\item{errors}{the variation imposed on the catches and the cpue, named
catchSD and cpueSD, both with default values = 1e-08, which do not
change the simulated values.}
}
\value{
histcatch, a matrix of year, catch-by-fleet, and cpue-by-fleet
}
\description{
samplefishery extracts the catches and the cpue from the
simulated fishery and can add error to both, as desired.
}
\examples{
\dontrun{
  library(codeutils)
  library(hplot)
  library(agestruct) # obviously setup your own direcdtory structure
  data(const2)
  glb <- initiateglobals(consts=const2,selabove=1)
  stock <- makestock(glb=glb)
  out <- simdynF(glb=glb,stk=stock,reps=8,full=TRUE)
  histC <- samplefishery(out,glb)
  print(histC)
}
# out=out;glb=glb;errors=c(catchSD=catchSD,cpueSD=cpueSD)
}
