% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{SBMctrltemplate}
\alias{SBMctrltemplate}
\title{SBMctrltemplate creates a draft control file for Size-Based Models}
\usage{
SBMctrltemplate(rundir, infile = "controlLBM.csv")
}
\arguments{
\item{rundir}{the full path to the working directory being used}

\item{infile}{what name will be given to the control .csv file}
}
\value{
nothing but it does generate a csv file into the rundir
}
\description{
SBMctrltemplate simplifies the generation of the working
control file for fitting a size-based model to population data. The
file must be a .csv file. We use Me instead of M to denote natural
mortality because M is not unique, which means R can become confused.
}
\examples{
\dontrun{
 rundir <- tempdir()
 SBMctrltemplate(rundir,infile="draftctrlLBM.csv")
 setup <- readSBMctrl(rundir,infile="draftctrlLBM.csv")
 # indir=rundir; filename="control2.csv"  #
}
}
\seealso{
{
\link{readSBMctrl}
}
}
