% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{SBMdatatemplate}
\alias{SBMdatatemplate}
\title{SBMdatatemplate creates a draft data file for LBM}
\usage{
SBMdatatemplate(rundir, outfile = "dataSBM.csv")
}
\arguments{
\item{rundir}{the full path to the working directory being used}

\item{outfile}{what name will be given to the data .csv file}
}
\value{
nothing but it does generate a csv file into the rundir
}
\description{
SBMdatatemplate simplifies the generation of the working
data file for fitting a size-based model to population data. The
file must be a .csv file. The data input includes the catches, the
cpue, any FIS data, and size-composition data.
}
\examples{
\dontrun{
 rundir <- tempdir()
 SBMdatatemplate(rundir,outfile="dataSBM.csv")
 setup <- readSBMdata(rundir,outfile="dataSBM.csv")
}
}
\seealso{
{
\link{readSBMdata}
}
}
