% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{SBMpintemplate}
\alias{SBMpintemplate}
\title{SBMpintemplate generates an example of a pin file's format}
\usage{
SBMpintemplate(rundir, infile, title = "A_rock_Lobster")
}
\arguments{
\item{rundir}{the full path to the working directory being used}

\item{infile}{the name to be given to the pin file}

\item{title}{text to be placed at the head of the pin file as a label}
}
\value{
nothing but it does generate a file named 'infile' within 'rundir'
}
\description{
SBMpintemplate is used to illustrate the required format of
the pin file. It may not hold examples of all parameters but it does
provide a template in the correct format that can be edited.
}
\examples{
print("wait on data")
}
\seealso{
{
\link{readSBMpin}
}
}
