% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agestruct.R
\name{setsigmaR}
\alias{setsigmaR}
\title{setsigmaR change recruitments variation when making data sets}
\usage{
setsigmaR(const, sigR = 0)
}
\arguments{
\item{const}{the output from 'getconstants'}

\item{sigR}{the recruitment variability}
}
\value{
the const object, perhaps with fishbiol$rec - sigmaR altered
}
\description{
setsigmaR is used when geberating simulated data sets and the
sigmaR derived from the the setup dayta file needs to be changed. If
sigR = 0 this does nothing except return the original
}
\examples{
data(simconst2)
print(simconst2$fishbiol$rec)
simconst3 <- setsigmaR(const=simconst2,sigR=0.2)
print(simconst3$fishbiol$rec)
}
