% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspm.r
\name{simbh}
\alias{simbh}
\title{simbh simulates the expected Beverton-Holt recruitment}
\usage{
simbh(spb, R0, B0, steep, sigmaR = 1e-09)
}
\arguments{
\item{spb}{the current spawning or mature biomass}

\item{R0}{the unfished average recruitment level}

\item{B0}{the unfished spawning biomass.}

\item{steep}{the steepness of the Beverton-Holt stock recruitment curve}

\item{sigmaR}{recruitment variability, a bias-corrected Log-Normal random
effect. Default = 1e-09. If < 0.001 then epsilon is set = 1.0 which
implies no effect.}
}
\value{
the expected Beverton-Holt recruitment level, a real number in the
linear scale
}
\description{
simbh simulates the expected Beverton-Holt stock recruitment
level from the available spawning biomass, the steepness, R0 and B0.
This would be used when simulating a stock recruitment relationship.
}
\examples{
simbh(10000,1500000,30000,0.75)                  # should be 1285714
simbh(spb=30000,R0=1500000,B0=30000,steep=0.75)  # should be 1500000
}
