% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{simconst2}
\alias{simconst2}
\title{simconst2 is the output of getconstants}
\format{
A list of siz objects
\describe{
\item{biology}{ a 1 clumn matrix of biological properties, eg M}
\item{fishbiol}{ a list of rec = c(B0, sigmaR, origsigR, steep), Rsplit,
which will be 1 if only one region, initdepl, and sigmaCE }
\item{fishery}{ a list of the selectivity parameters for each fleet}
\item{strucure}{details the structural form of the model, years, ages,
sizes, fleets, selecttype}
\item{histC}{a 45 year x 3 matrix of twl catches, auln catches, and year}
\item{randseed}{ a random seed to allow for repeatability}
}
}
\usage{
simconst2
}
\description{
simconst2 is a list contianing matrices, lists, and a scalar
value. It is the outptu from teh getconstants function. See the
description in the Simulation MOdel Appendix in \emph{More Fisheries
Modelling using R}.
}
\examples{
 data(simconst2)
 str(simconst2)
}
\keyword{datasets}
