% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sizecomp}
\alias{sizecomp}
\title{sizecomp a 38 x 20 matrix of sizecomposition data for abalone}
\format{
A 38 x 20 matrix of integer counts
\describe{
\item{rows}{ 38 2 mm size-classes from 136 - 210 mm}
\item{columns}{ 20 years from 1993 - 2020}
}
}
\usage{
sizecomp
}
\description{
sizecomp a a 38 x 20 matrix of 2mm size-classes from 136 - 210
mm and 20 years from 1993 - 2020, with counts in each size-class in each
year. Actual sample sizes vary from 345 - 9960, although effective
sample size, of course, is lower than these limits.
}
\examples{
data(sizecomp)
sizecomp
}
\references{
Haddon, M. (2025) aMSE: A Software Framework for Abalone
Management Strategy Evaluation_. R package version 1.0.5,
\url{https://github.com/haddonm/aMSE}.
}
\keyword{datasets}
