% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{STMvB}
\alias{STMvB}
\title{STMvB Generates a Size Transition Matrix for von Bertalanffy growth}
\usage{
STMvB(p, mids)
}
\arguments{
\item{p}{a vector of three parameters in the following order
Linf the maximum size of all animals, K the growth rate parameter that
determines how quickly each individual approaches the maximum size, and
growCV, the coefficient of variation of the growth such that it
describes the standard deviation of any Normal distribution used to
describe the distribution or spread of sizes resulting from a given
initial size. It does this through the sd = expected mean size (initial
size + mean increment) multiplied by the growCV (sd = x_bar x growCV).}

\item{mids}{a vector of initial lengths which also define the width of
each size class thus, mids from 1 - 60 would imply size classes 1, 2, 3,
... 60. With the units matching the original measurements.}
}
\value{
A square matrix with dimension = the length of the mids vector
}
\description{
STMvB With the input of the three parameters inside a vector,
and a vector of initial lengths or mid-points of size classes STMvB
generates a square transition matrix with the probabilities of
growing from each initial size into the same or larger sizes.
Negative growth is disallowed so the maximum size class = Linf. All
columns in the matrix sum to one.The von Bertalanffy growth curve
implies that the growth increments decline linearly with initial size.
This, in turn implies that when an animal reaches the Linf it will stay
at that size. This uses the Fabens version of the vB curve, which has
different implications to the age-based vB curve. The implementation
here does not have a plus-group in the final size. When the curve
implies there would be negative growth the cut-off means that initially
some of the later columns do not sum to 1.0. Rather than lumping all the
'missing' fish into the largest size class the total for each column is
used to rescale the proportion of fish in each of the size classes into
which it grows.
}
\examples{
 pars <- c(Linf=75.0,K=0.275,CVvb=0.08)
 lens <- seq(3,73,5)
 M <- 0.25; surv <- exp(-M)
 G <- STMvB(p=pars,mids=lens)
 print(round(G,4))
 nyr <- 20
 Nt <- matrix(0,nrow=15,ncol=nyr,dimnames=list(lens,2000:(1999+nyr)))
 rec <- numeric(15); rec[1] <- 100
 Nt[,1] <- rec
 for (i in 1:(nyr-1)) Nt[,i+1] <- G \%*\% (surv * Nt[,i]) + rec
 round(Nt,1)
}
\references{
Haddon, M. (2011) \emph{Modelling and Quantitative Methods in
Fisheries} 2nd edition. Boca Raton: Chapman & Hall/CRC. 449 pp.
ISBN: 978-1-58488-561-0

Fabens, A.J. (1965) Properties and fitting of the von Bertalanffy
growth curve. \emph{Growth}, \emph{29}: 265-289.
}
