% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agestruct.R
\name{template2F1S}
\alias{template2F1S}
\title{template2F1S data template for a 2 fleet 1 region/stock fishery}
\usage{
template2F1S(rundir, filename = "F2S1.csv")
}
\arguments{
\item{rundir}{directory in which to find the data file and run the analysis}

\item{filename}{the name of the data file to be produced, default=
'F21S.csv'.}
}
\value{
the function write a data file to rundir and returns the filename
}
\description{
template2F1S data template for a 2 fleet 1 region/stock fishery
}
\examples{
\dontrun{
  rundir <- tempdir()
  dirExists(rundir,verbose=TRUE)
  template2F1S(rundir,filename="Flt2Stock1.csv")
  dir(rundir)
}
}
