% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{TRL}
\alias{TRL}
\title{TRL a list of five objects realting to an hypothetical lobster fishery}
\format{
A list of five objects, both matrices and lists
\describe{
\item{fish}{34 years x 4 (year, LML, catch, and cpue) matrix 1988 - 2021}
\item{sizecomp}{31 size-class x 17 year x 2 sexes of size-composition data}
\item{glb}{a globals object containing model constants such as natural
mortality, size-classes, steepness, etc. see str(glb)}
\item{constants}{14 constants relating to weight-at, maturity-at, and
growth-at-size for males and females}
\item{allpin}{initial parameter values prior to fitting, 27 x 2 matrix,
with the second column determining which parameters are fitted}
}
}
\usage{
TRL
}
\description{
TRL is a list of five objects relating to an hypothetical
tropical rock lobster fishery. It contains fish, sizecomp, glb,
constants, and allpin. These can be used to generate a stock of rock
lobster for use in a size-based model (SBM). The males and females have
different growth dynamics and so require separate STM.
}
\examples{
data(TRL)
str(TRL)
}
\keyword{datasets}
