% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agestruct.R
\name{unfished}
\alias{unfished}
\title{unfished - generates the numbers at age for an unfished population}
\usage{
unfished(glb)
}
\arguments{
\item{glb}{\itemize{
\item the global constants object containing biology, fishery and
structure
}}
}
\value{
a list containing N0, R0, A0, SpB0, and expB0
}
\description{
unfished - generates the numbers at age for an unfished
population, and determines the recruitment dynamics
}
\examples{
\dontrun{
library(fmr)
datafileTemplate(filename="C:/A_CSIRO/Rcode/agestructrun/constants.csv")
const <- getconstants("C:/A_CSIRO/Rcode/agestructrun/constants.csv")
globals <- initiateglobals(const)
unfish <- unfished(globals)
}
}
