% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{vbfabens}
\alias{vbfabens}
\title{vbfabens calculates predicted mean growth increments for vB Fabens}
\usage{
vbfabens(p, mids)
}
\arguments{
\item{p}{a vector of two vB parameters Linfinity, the maximum size, and K
the rate of growth towards the maximum size.}

\item{mids}{the center of each size class for which the mean expected
groeth increment will be calculated.}
}
\value{
a vector of growth increments the samelength as the vector of mid-
points.
}
\description{
vbfabens calculate the predicted mean growth increment
expected from the Fabens version of the von Bertalanffy growth curve.
This is designed to be fitted to tagging data which is used to estimate
the mean increment of different sized animals. Assuming vB growth
implies a linear decrease in the growth increment with increasing size.
}
\examples{
 mids <- seq(3,75,5)
 p <- c(75,0.26)
 Linc <- vbfabens(p,mids)
 print(cbind(mids,Linc))
}
