% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{westroughy}
\alias{westroughy}
\title{westroughy Three data objects describing western Tasmanian Orange Roughy.}
\format{
A list of three objects
\describe{
\item{fish}{ a data.frame containing Year, Catch, standardized CPUE, and
SE, the standard error of the CPUE estimates, if present}
\item{glb}{ a list of global variables including maxage, M, parameters for
growth, weight-at-age, maturity-at-age, steepness, R0, selectivity,
resilience, number of ages, and the ages themselves. }
\item{props}{ a data.frame of age, laa, waa, maa, sela, and feca}
}
}
\usage{
westroughy
}
\description{
westroughy a dataset containing the fish data.frame, the glb
list, and the props data.frame set up ready for analysis using an
age-structured production model. The analysis of the CPUE, not typically
available for orange roughy fisheries is described in the reference
below.
}
\examples{
data(westroughy)
fish <- westroughy$fish
glb <- westroughy$glb
props <- westroughy$props
str(glb)
}
\references{
Haddon, M. (2018) Western Orange Roughy. pp806-821 \emph{in}
Tuck, G.N. (ed.) \emph{Stock Assessment for the Southern and Eastern
Scalefish and Shark Fishery 2016 and 2017. Part 2, 2017}. Australian
Fisheries Management Authority and CSIRO Oceans and Atmosphere, Hobart.
837p.
}
\keyword{datasets}
