% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspm.r
\name{aspmindexfit}
\alias{aspmindexfit}
\title{aspmindexfit plots an aspm fit to CPUE or Index data}
\usage{
aspmindexfit(infish, glb, CI = NULL, rundir = "", console = TRUE)
}
\arguments{
\item{infish}{the matrix of fishery dynamics from the dynamicsH, dynamicsF,
of dynF functions}

\item{glb}{the globals object for the aspm analysis used}

\item{CI}{the output of the getLNCI function that defines Log-Normal CI}

\item{rundir}{default = "", give a full path if saving a png file}

\item{console}{should the plot be sent to the console, default = TRUE. If
set = FALSE, then rundir can be set to identify where the png of the
plot will be saved}
}
\value{
nothing but it does generate a plot
}
\description{
aspmindexfit plots an aspm fit to CPUE or Index data with the
option of including 95\% Log-Normal CI around the model fit.
}
\examples{
data("westroughy")
fish <- westroughy$fish; glb <- westroughy$glb
props <- westroughy$props
pars <- c(7.1,-1,-7.7)
ans <- fitASPM(pars,dynamicsH,infish=fish,inglb=glb,inprops =props)
out <- dynamicsH(ans$estimate,infish=fish,inglb=glb,inprops=props,
                 full=TRUE)
ceCI <- getLNCI(out$fishery[,"predCE"],exp(ans$estimate[2]))
aspmindexfit(infish=out$fishery,glb=glb,CI=ceCI,console=TRUE)
}
\seealso{
{
\link{getLNCI}, \link{dynamicsH}, \link{dynamicsF}, \link{dynF}
}
}
