% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmr_utils.R
\name{bh}
\alias{bh}
\title{bh calculates the Beverton-Holt recruitment}
\usage{
bh(spb, R0, B0, steep, devR = 1)
}
\arguments{
\item{spb}{the current spawning biomass}

\item{R0}{the unfished recruitment levels for the population}

\item{B0}{the unfished spawning biomass; scalar}

\item{steep}{the steepness for the population; scalar}

\item{devR}{recruitment deviate for a given year. default = 1.0, meaning
the value will come off the stock recruitment curve.}
}
\value{
an absolute number of recruits from a given spawning biomass
}
\description{
bh calculates the Beverton-Holt recruitment for the
input populations in a single year; parameterized with steepness,
R0, and B0. This is for use when fitting a model as it allows for the
direct input of recruitment deviates.
}
\examples{
print("wait on data")
}
