% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agestruct.R
\name{bhsim}
\alias{bhsim}
\title{bhsim - calculates number of bh from input spawning biomass}
\usage{
bhsim(inSB, R0, B0, steep, sigmaR)
}
\arguments{
\item{inSB}{input spawning biomass in tonnes.}

\item{R0}{unfished recruitment levels}

\item{B0}{unfished spawning biomass}

\item{steep}{the steepness of the Beverton-Holt SR relationship}

\item{sigmaR}{recruitment variability}
}
\value{
a scaler representing the number of recruits
}
\description{
bhsim - calculates number of recruits from input spawning
biomass assumes steep, R0, B0, and sigmaR are available. It
also assumes the use of the Beverton - Holt stock recruitment curve.
Only used in simdynF
}
\examples{
\dontrun{
 library(fmr)
 data(const)
 globals <- initiateglobals(const)
 pop <- makestock(glb=globals)
 bh(2500,pop$R0,pop$B0,sigmaR=1e-07,globals$steep) # = 1763871
} 
}
