% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.R
\name{comparevars}
\alias{comparevars}
\title{comparevars plots a given variables changes between two scenarios}
\usage{
comparevars(
  yrs,
  var1,
  var2,
  varname,
  scenarios,
  console = TRUE,
  rundir = "",
  prepplot = FALSE,
  legpos = "topright",
  cex = 1
)
}
\arguments{
\item{yrs}{the years of the fishery dynamics}

\item{var1}{a variable from the the dynamics of the first scenario}

\item{var2}{the same variable from teh second scenario}

\item{varname}{The name of the variable being compared}

\item{scenarios}{a character vector of the names of the two scenarios}

\item{console}{should the plot be printed to the console or a file?
default = TRUE}

\item{rundir}{the full path of the rundir where analyses are occurring.
defauklt = ''}

\item{prepplot}{default = FALSe, implying the plot will be part of a
composite whose structure is defined elsewhere. If TRUE, then a single
comparison will be produced.}

\item{legpos}{default = 'topright' the position of the legend, which is
simply the contents of the scenarios argument.}

\item{cex}{default = 1, font size for single plots, ie prepplot=TRUE}
}
\value{
if console = FALSe it returns a filename and a file is generated,
if console = TRUE, nothing is returned but it does generate a plot
}
\description{
comparevars can be used when comparing the dynamics of two
alternative scenarios when fitting an integrated assessment model. It
requires the fishery dynamics of each scenario and plots whichever
variables are being considered (see example).
}
\examples{
\dontrun{  # illustrate typical use
  fish49 <- opt49$fishery
  fish45 <- opt45$fishery
  yrs <- fish49[,"year"]
  plotprep(width=10,height=8)
  parset(plots=c(2,2),cex=1)
  comparevars(yrs=yrs,var1=fish49[,"deplete"],var2=fish45[,"deplete"],
              varname="Spawning Biomass Depletion",
              scenarios=c("fish49","fish45"),console=TRUE,prepplot=FALSE)
  comparevars(yrs=yrs,var1=fish49[,"twlPCE"],var2=fish45[,"twlPCE"],
              varname="Predicted CPUE",
              scenarios=c("fish49","fish45"),console=TRUE,prepplot=FALSE)
  comparevars(yrs=yrs,var1=fish49[,"spawnB"],var2=fish45[,"spawnB"],
              varname="Predicted Spawning Biomass",
              scenarios=c("fish49","fish45"),console=TRUE,prepplot=FALSE)
  comparevars(yrs=yrs,var1=fish49[,"recruit"],var2=fish45[,"recruit"],
              varname="Predicted Recruitment",scenarios=c("fish49","fish45"),
              console=TRUE,prepplot=FALSE)
}
}
