% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspm.r
\name{dynamicsH}
\alias{dynamicsH}
\title{dynamicsH describe the ASPM dynamics using annual harvest rates}
\usage{
dynamicsH(
  pars,
  infish,
  inglb,
  inprops,
  waa = "waa",
  maa = "maa",
  sela = "sela",
  full = FALSE
)
}
\arguments{
\item{pars}{the dynamics relies on many parameters sitting in the function's
environment, these are ages, nages, maxage, M, maa, waa, sela, fish,
and nyrs. 'pars' can contain either two to three log-transformed
parameters: 1) is the log of average unfished recruitment, inR0. 2) is
the variability around the index of relative abundance (cpue), not used
directly in the dynamics but rather in the estimation of the likelihoods
during the fitting process, 3) if is present, is the catchability 'q',
which alternatively can be estimated using the closed form. 4) If
present this would be the log of the initail depletion.}

\item{infish}{the fish data.frame from readdata or an internal dataset}

\item{inglb}{the glb data.frame from readdata or an internal dataset}

\item{inprops}{the props data.frame from readdata or an internal dataset}

\item{waa}{the character name of the weight-at-age}

\item{maa}{the character name of the maturity-at-age}

\item{sela}{the character name of the selectivity-at-age}

\item{full}{should all outputs from dynamics be given. When fitting the
model, set this to FALSE. Once fitted, change this to TRUE to get all
the required outputs.}
}
\value{
a data.frame containing the fishery dynamics according to the input
parameter inR0. In particular it includes the Catch and PredC, and the
CPUE and PredCE, which can be used in a maximum likelihood context.
}
\description{
dynamicsH summarizes the dynamics of the Age-Structured
Production Model (ASPM) in which the catches are represented as annual
harvest rates. Fitting the ASPM entails estimating the unfished
recruitment level (R0), which is input as a parameter. There may be
other parameters as described in the pars section.
}
\examples{
data(dataspm)
fish <- dataspm$fish
glb <- dataspm$glb
props <- dataspm$props
par <- c(glb$R0,0.20)  # not fitted to the data, this is just an initial guess
fishery <- dynamicsH(par,infish=fish,inglb=glb,inprops=props,full=FALSE)
print(fishery)
# pars=bestL$pars;infish=fish;inglb=glb;inprops=props; full=TRUE
}
\seealso{
{
\link{dynamicsF}
}
}
