% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspm.r
\name{dynF2}
\alias{dynF2}
\title{dynF2 describe the ASPM dynamics for a two-fleet model}
\usage{
dynF2(
  pars,
  infish,
  inglb,
  inprops,
  waa = "waa",
  maa = "maa",
  sela = fleets,
  full = FALSE,
  reps = 6
)
}
\arguments{
\item{pars}{the initial parameter values lnR0, lnq1, lnq2}

\item{infish}{the fishery statistics as in year, catches, CPUE, year, twl,
auln, twlCE, aulnCE, with NA instead of zeros, from readASPMdata}

\item{inglb}{the glb data.frame from readASPMdata or built in dataset const2.}

\item{inprops}{matrix of properties, age, then length-, weight-, maturity,
and selectivity-at-age for each fleet, from readASPMdata}

\item{waa}{name used for weight-at-age in inprops, default='waa'}

\item{maa}{name used for maturity-at-age in inprops, default='waa'}

\item{sela}{name used for maturity-at-age in inprops,a vector of the
names of each fleet, usually=glb$fleets}

\item{full}{should all outputs from dynamics be given. When fitting the
model, set this to FALSE. Once fitted, change this to TRUE to get all
the required outputs.}

\item{reps}{how many iterations to use in findFs, default = 6}
}
\value{
if !full then a data.frame containing the fishery dynamics. This
includes year, catch-by-fleet, PredC-by-fleet, cpue-by-fleet,
predCE-bt-fleet, start of year exploitB-by-fleet, total spawning
biomass, spawning biomass depletion, recruitment, predictedF-by-fleet.
}
\description{
dynF2 summarizes the dynamics of an Age-Structured
Integrated Production Model (ASPM). Fishing mortality is implemented as
instantaneous rates. Nt is the numbers-at-age at
the start (or end) of a year, catchN is the predicted numbers-at-age in
the catch of each fleet, NumC is the total numbers-at-age in the catch,
Lt is the predicted numbers-at-size at the start of each year, Gtran\emph{Nt,
LCflt is the numbers-at-size in the catch for each fleet, Gtran}catchN,
at the end of each year, and LC is the total numbers-at-size in the
catch, which is Gtran * NumC. These numbers-at-size ignore the fact that
larger fish are differentially taken by size due to selectivity (a weak
assumption in purely age-structured models.
}
\examples{
\dontrun{
 library(fmr)
 library(codeutils)
 library(hplot)
 rundir <- "c:/run2F1S/" # OBVIOUSLY DEFINE YOUR OWN DIRRECTORY STRUCTURE
 dirExists(rundir,verbose=FALSE)  # this will create rundir if required
 always be very careful when using R to manipulate files and directories
 template2F1S(rundir,filename="test2F1S.csv")  # generates test2F1S.csv 
 const2 <- readASPMdata(pathtopath(rundir,"test2F1S.csv"))
 fish <- const2$fish
 glb <- const2$glb
 props <- makeprops(const2,selabove=1)
 pars <- log(c(R0=1825000, qc1=0.0003, qc2=0.00088))
 outdyn <- dynF2(pars=pars,infish=fish,inglb=glb,inprops=props,
                 waa="waa",maa="maa",sela=glb$fleets,full=FALSE,reps=6) 
 print(outdyn)  # should be ~ 26.8              
 outdyn <- dynF2(pars=pars,infish=fish,inglb=glb,inprops=props,
                 waa="waa",maa="maa",sela=glb$fleets,full=TRUE,reps=6) 
 str(outdyn)
}
}
\seealso{
{
\link{template2F1S}, \link{readASPMdata}
}
}
