% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspm.r
\name{findF}
\alias{findF}
\title{findF uses an iterative strategy for finding each year's F value}
\usage{
findF(cyr, Nyr, sel, aaw, M, reps = 8)
}
\arguments{
\item{cyr}{the known catch in a given year}

\item{Nyr}{the numbers at size at the start of the given year or end of the
year before}

\item{sel}{the selectivity of the fishing gear}

\item{aaw}{the weight-at-age}

\item{M}{the instantaneous natural mortality rate}

\item{reps}{how many internal loops to use finding each F, default = 6}
}
\value{
the fully selected fishing mortality rate
}
\description{
findF is used when searching for the instantaneous F that will
produce the required catch in a given year in an age-structured model. It
needs generalizing to operate with multiple gears.
}
\examples{
print("wait on example data sets")
}
