% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{findFs}
\alias{findFs}
\title{findFs uses an iterative strategy for finding each year's F values}
\usage{
findFs(cyr, Nyr, sel, aaw, M, reps = 8)
}
\arguments{
\item{cyr}{a vector of known catches in a given year for nfleets}

\item{Nyr}{the numbers at size at the start of the given year or end of the
year before}

\item{sel}{a matrix of the selectivity of the fishing gears}

\item{aaw}{the weight-at-age}

\item{M}{the instantaneous natural mortality rate}

\item{reps}{how many internal loops to use finding each F, default = 6}
}
\value{
the fully selected fishing mortality rates
}
\description{
findFs is used when searching for the instantaneous F that will
produce the required catch in a given year in an age-structured model
when there are multiple gears, with different selectivity being used.
}
\examples{
print("wait on example data sets")
}
