% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspm.r
\name{fitASPM}
\alias{fitASPM}
\title{fitASPM fits an age-structured production model}
\usage{
fitASPM(initpar, minfun, infish, inglb, inprops, hessian = FALSE)
}
\arguments{
\item{initpar}{a vector of 2 or 3 numbers that are the initial parameter
values given to the estimate of logR0, and the estimate of the variation
around the CPUE data that the model is to be fitted to, and finally, the
catchability coefficient.}

\item{minfun}{the dynamics function outputning a -ve log-likelihood}

\item{infish}{the fish data.frame from readdata or built in dataset}

\item{inglb}{the glb data.frame from readdata or built in dataset}

\item{inprops}{the props data.frame from readdata or built in dataset}

\item{hessian}{should the hessian be estimated, default = FALSE}
}
\value{
a list containing the optimal output
}
\description{
fitASPM fits an age-structured production model that can have
up to three parameters, R0 the unfished recruitment level, se the
variation around the estimated CPUE, and avq the catchability
coefficient. The inverse Hessian, when using log-likelihoods, provides
access to the variance-covariance matrix, which in turn can be used
to generate standard errors on the model parameters. This is used when
using asypmtotic errors to caharacterize uncertainty. See section 6.5
in Haddon (2021).
}
\examples{
data("westroughy")
fish <- westroughy$fish
glb <- westroughy$glb
props <- westroughy$props
pars <- c(7,0.3,-7.7)
bestL <- fitASPM(pars,dynamicsH,infish=fish,inglb=glb,inprops=props,
                 hessian=TRUE)
bestL
out <- dynamicsH(bestL$estimate,fish,glb,props,full=TRUE)
round(out$fishery,4)
}
\references{
Haddon, M. (2021) Using R for Modelling and Quantitative
Methods in Fisheries, CRC Press / Chapman & Hall/ Boca Raton 337p.
ISBN: 9780367469894
}
