% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_tables.R
\name{fm_survey}
\alias{fm_survey}
\title{The survey main table}
\usage{
fm_survey(key, std = TRUE, trim = TRUE)
}
\arguments{
\item{key}{your FM API key}

\item{std}{boolean, if TRUE (default) variables are renamed according to
{fmr} convention}

\item{trim}{boolean, if TRUE (defaul) only essential (standard) variables
are returned. Only active if arguement std is set to TRUE.}
}
\value{
a tibble
}
\description{
The function returns information about each survey.
}
