% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_tables.R
\name{fm_survey}
\alias{fm_survey}
\title{The survey main table}
\format{
A data frame with nine variables:
\describe{
\item{\code{site}}{Name of the (landing) site}
\item{\code{island}}{Name of the island, returned in some cases}
\item{\code{date}}{Date of the survey}
\item{\code{time}}{Time of the survey}
\item{\code{survey_id}}{Numerical code, used to join to other tables}
}

For further details, see \url{https://fimsehf.atlassian.net/wiki/spaces/FT/pages/3283812558/New+The+Data+Structure+Explained}
}
\usage{
fm_survey(key, std = TRUE, trim = TRUE)
}
\arguments{
\item{key}{your FM API key}

\item{std}{boolean, if TRUE (default) variables are renamed according to
{fmr} convention}

\item{trim}{boolean, if TRUE (defaul) only essential (standard) variables
are returned. Only active if arguement std is set to TRUE.}
}
\value{
a tibble
}
\description{
The function returns information about each survey.
}
