% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_connection.R
\name{fm_tbl}
\alias{fm_tbl}
\title{Fisheries manager table connection}
\usage{
fm_tbl(table = "siteD", key, clean = TRUE)
}
\arguments{
\item{table}{string, see what is currently available at: https://fimsehf.atlassian.net/wiki/spaces/FT/pages/3255894017/.REST+API+for+reporting+endpoint}

\item{key}{string}
}
\value{
a tibble
}
\description{
Fisheries manager table connection
}
