% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_table.R
\name{fm_tbl}
\alias{fm_tbl}
\title{Fisheries manager table connection}
\usage{
fm_tbl(table = "siteD", key)
}
\arguments{
\item{table}{string}

\item{key}{string}
}
\value{
a tibble
}
\description{
Fisheries manager table connection
}
