% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agestruct.R
\name{getconstants}
\alias{getconstants}
\title{getconstants - reads the input constants that condition the model}
\usage{
getconstants(infile = "constants.csv")
}
\arguments{
\item{infile}{\itemize{
\item defaults to 'constants.csv'. Needs to be a csv file
}}
}
\value{
a list of vectors 'biology' and 'fish' plus a lists of 'fishbiol',
'fishery', and 'glb'
}
\description{
getconstants - reads the input constants that condition the
model. The constants include all those produced by the function
constantfileTemplate: M, surv, Linf, K, t0, growCV, Wta, Wtb, steep,
age50M, deltaM, B0, sigmaR, etc. The data are read in from 'infile' and
must contain sections entitled: BIOLOGY, FISHERY, STRUCTURE, and
HISTORICALCATCH
}
\examples{
\dontrun{
  rundir <- tempdir()
  template2F1S(rundir,filename="Flt2Stock1.csv")
  const2 <- getconstants(infile=paste0(rundir,"//Flt2Stock1.csv"))
  str(const2)
}
}
