% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agestruct.R
\name{getfltcatchsamp}
\alias{getfltcatchsamp}
\title{getfltcatchsamp makes bootstrap samples of length and age composition}
\usage{
getfltcatchsamp(
  fltcatchN,
  nyrs,
  n = 500,
  randsd = 40,
  maxsamp = 2000,
  sampvar = 50
)
}
\arguments{
\item{fltcatchN}{the predicted catch as numbers-at-age out$catchN<,,flt> or
numbers-at-length LCflt<,,flt> from the out list produced by simdynF.}

\item{nyrs}{the total number of years in the dynamics produced by simdynF.}

\item{n}{the expected sample size}

\item{randsd}{the variation in sample size used by rowfreqboot}

\item{maxsamp}{the maximum sample size possible (> n)}

\item{sampvar}{the variation in how much each catch numbers are reduced
prior to bootstrapping.}
}
\value{
an invisible matrix of nages or nsizes x nyrs, of counts by age or
length by year
}
\description{
getfltcatchsamp is used to generate bootstrap samples from
given years of age-composition or length-composition data out of
simdynF. By default is samples for all years in which composition data
are available. If samples are assumed to be taken less frequently, it is
up to teh user to thin the output. The output data (and input data) can
be visualized using plotcompdata from codeutils. getfltcatchsamp uses
the rowfreqboot function from codeutils. The function works on the
composition data (age or length) for a single fleet. The function first
reduces the number of observed composition numbers in each year as the
originals can be 100s of thousands, which would slow the bootstrap
process.
}
\examples{
print("wait on data sets")
}
