% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{IAdynF}
\alias{IAdynF}
\title{IAdynF the dynamics of an IA model with CPUE and age-composition data}
\usage{
IAdynF(
  pars,
  fish,
  glb,
  props,
  agecomp,
  waa = "waa",
  maa = "maa",
  fleet = "twl",
  year = "year",
  catch = "twl",
  cpue = "twlCE",
  full = FALSE,
  reps = 6
)
}
\arguments{
\item{pars}{the dynamics relies on many variables being avilable in
particular ages, nages, maxage, M, fish, nyrs, and maa, waa, which are
contained in props. As selectivity is being estimated it is recalculated
each iteration of the fitting process using the new parameter set. In an
integrated assessment the intpu parameters, 'pars', contains multiple
parameters, each of which is log-transformed so that all parameters are
close to the same magnitude: 1) is the log-transformed average unfished
recruitment, inR0. 2) is the catchability 'q' for each fishing gear 3)
the selectivity parameters for each gear type, and finally, 4) the
recruitment deviates selected for estimation.}

\item{fish}{the fishery data, an object out of const}

\item{glb}{the globals object, out of const}

\item{props}{the properties of waa, maa, and laa, out of const}

\item{agecomp}{the age-composition of catches, out of const}

\item{waa}{the character name of the weight-at-age in props}

\item{maa}{the character name of the maturity-at-age  in props}

\item{fleet}{the character name of the active fleet, default = 'twl', used
to identify the agecomposition data in teh all agecomp list.}

\item{year}{the character name of the year column in fish, default='year'}

\item{catch}{the character name of the catch column in fish, default='twl'}

\item{cpue}{the character name of the cpue column in fish, default='twlCE'}

\item{full}{should all outputs from dynamics be given. When fitting the
model, set this to FALSE. Once fitted, change this to TRUE to get all
the required outputs.}

\item{reps}{how many loops within the findF function to find each F}
}
\value{
a data.frame containing the fishery dynamics according to the input
parameter inR0. Includes Year, Catch, PredC, SpawnB, ExploitB, FullH,
CPUE, PredCE, Deplete, Recruit, FullF. If full=TRUE then a list with all
dynamic outputs.
}
\description{
IAdynF summarizes the dynamics of an Integrated Stock
Assessment model that includes both CPUE and age-composition data.
Fitting this model entails estimating the unfished recruitment level
(R0), which is input as a parameter, along with, for each gear,
catchability and selectivity. Selectivity-at-age parameters can only be
estimated when there are age-composition data available. Finally, if
sufficient age-composition data are available across years it becomes
possible to estimate recruitment deviates through time. Once again we
have included a penalty on catches to ensure that predC = ObsC, and a
penalty on the recruitment devaites to avoid over-fitting to data.
}
\examples{
\dontrun{
data(westroughy)
fish <- westroughy$fish
glb <- westroughy$glb
props <- westroughy$props
pars <- c(7.1,-1,-7.7) # logR0, sigCE, estimate avq
fishery <- dynF(pars,fish,glb,props) 
pars <- c(7.1,-1,-7.7)
bestL <- optim(pars,dynF,method="Nelder-Mead",fish=fish,glb=glb,
               inprops=props,control=list(maxit=1000,parscale = c(10,1,10)))
str(bestL)
print(bestL)
out <- IAdynF(bestL$par,fish,glb,props,full=TRUE)
print(round(out$fishery,4)) 
}
# pars=pars;fish=fish;glb=glb;props=props;agecomp=allagecomp;
# waa="waa";maa="maa";fleet="twl";year="year";
# catch="twl";cpue="twlCE";full=TRUE;reps=6
}
\seealso{
{
\link{dynamicsF}, \link{dynamicsF}, \link{findF}, \link{findFs}
}
}
