% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmr_utils.R
\name{logistic}
\alias{logistic}
\title{logistic standard selectivity function}
\usage{
logistic(L50, delta, depend)
}
\arguments{
\item{L50}{is the length/age at 50 percent selection/maturity/whatever}

\item{delta}{is l95 - l50 selection/maturity/whatever, nominalscale}

\item{depend}{a vector of lengths/ages for which the logistic value will be
calculated.}
}
\value{
A vector of length/age(depend) containing predicted logistic values
}
\description{
logistic calculates a Logistic curve that can be used as a
selectivity function, or maturity curve, of wherever a logistic is
required. This version uses the logistic function
1/(1+exp(-log(19.0)*(lens-inL50)/(inL95-inL50))),
which explicitly defines the SM50 and uses SM95 as the second parameter.
}
\examples{
in50 <- 3
delta <- 1.25
ages <- seq(1,10,1)
select <- logistic(L50=in50,delta=delta,depend=ages)
round(cbind(ages,select),5)
}
