% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agestruct.R
\name{makedataset}
\alias{makedataset}
\title{makedataset generates a csv file from siumulated data}
\usage{
makedataset(
  const,
  glb,
  out,
  filename,
  rundir = "",
  datatitle = "",
  catchSD = 1e-08,
  cpueSD = 1e-08
)
}
\arguments{
\item{const}{the output object from getconstants}

\item{glb}{the output object from initiateglobals}

\item{out}{the output object from simdynF}

\item{filename}{the name of the datafile, MUST be a '.csv' file}

\item{rundir}{the directory path to the subdirectory in which all analysis
data and results are stored, default = ''}

\item{datatitle}{line 1 of data file decribing its contents, default = ''}

\item{catchSD}{standard deviation of normal variation imposed on reported
catches. default = 1e-08, effectively no variation}

\item{cpueSD}{standard deviation of normal variation imposed on reported
CPUE. default = 1e-08, effectively no variation}
}
\value{
nothing but it does write a data file to rundir/filename
}
\description{
makedataset generates a csv file from simulated data produced
by getconstants, initiateglobals, and simdynF.
}
\examples{
\dontrun{
 const2 <- getconstants(pathtopath(datraw,"F2-A1-S1.csv"))
 glb <- initiateglobals(consts=const2,selabove=1)
 stock <- makestock(glb=glb)
 out <- simdynF(glb=glb,stk=stock,reps=5,full=TRUE)
 catchSD=1e-08
 cpueSD=1e-08
 sigR=1e-08
 filename="F21S.csv"
 datatitle <- "A two-fleet one stock/region fishery data set"
 makedataset(const2,glb,out,filename,rundir='',datatitle="test",
             catchSD=1e-08,cpueSD=1e-08,sigR=1e-08)
 # now open 'F21S.csv
}
}
\seealso{
{
\link{getconstants}, \link{initiateglobals}. \link{simdynF}
}
}
