% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agestruct.R
\name{makeprops}
\alias{makeprops}
\title{makeprops generates a matrix of biological and selectivity properties}
\usage{
makeprops(const, selabove = 1)
}
\arguments{
\item{const}{a structured data file in the format defined by template2F1S}

\item{selabove}{a vector of ages or sizes that are never selected. Identify
their indices in the selecitivyt vector here. eg. if age 0 is never
selected then set zerocomps = 1, if ages 1-2 then zerocomps = c(1,2).}
}
\value{
a matrix of biological and selectivity propoerties
}
\description{
makeprops is a wrapper function that uses a constants file
to generate a matrix of the constant biological properties for a fishery
including length-at-age, weight-at-age, maturity-at-age, and the
selectivity of each fleet
}
\examples{
\dontrun{
  rundir <- tempdir()
  dirExists(rundir,verbose=TRUE)
  template2F1S(rundir,filename="test2F1S.csv")
  const2 <- getconstants(pathtopath(rundir,"test2F1S.csv"))
  fish <- const2$histC
  glb <- const2$structure
  props <- makeprops(const2,selabove=1)
  print(props)
}
}
\seealso{
{
\link{template2F1S}, \link{calcsel},
}
}
