% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmr_utils.R
\name{matchC}
\alias{matchC}
\title{matchC gives difference between predicted catch and observed catch}
\usage{
matchC(f, M, cyr, Byr)
}
\arguments{
\item{f}{the trial predicted instantaneous fishing mortality}

\item{M}{the instantaneous natural mortality value}

\item{cyr}{the catch in yr t}

\item{Byr}{the exploitable biomass at the start of yr t or end of t-1}
}
\value{
the absolute difference between the predicted and observed catches
}
\description{
matchC for a given instantaneous fishing mortality rate
calculates the difference between the predicted and observed catches.
This is to be used by the optimize function which uses 'a combination of
a golden section search and successive parabolic interpolation' to
minimize the difference between the absolute difference between the
catch and the predicted catch.
}
\examples{
matchC(f=0.04,M=0.05,cyr=15340,Byr=397388.66)
matchC(f=0.04037,M=0.05,cyr=15340,Byr=397388.66)
out <- optimize(matchC,interval=c(0,1),M=0.05,cyr=15340,Byr=397388.66)
out
f <- out$minimum
(397388.66 * (1 - exp(-(0.05 + f))) * f/(0.05 + f))
}
