% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{outlike}
\alias{outlike}
\title{outlike a 51 x 51 matrix of likelihoods for LnR0 and Lnq}
\format{
A 51 x 51 data.frame
}
\usage{
outlike
}
\description{
A 51 x 51 data.frame with rows as LnR0 = seq(5.9,8.4,0.05) and
columns of Lnq = seq(-8.7,-6.2,0.05). The values are the -veLL for each
combination when combined with a constant LnsigCE = -1.3. This can also
be used with persp to generate a 3D plot of the likelihood surface.
Note, in teh example,
}
\examples{
data(outlike)
p1 <- seq(5.9,8.4,0.05)
p2 <- -1.3
p3 <- seq(-8.7,-6.2,0.05)
z <- outlike
x <- p1
y <- p3
trans <- persp(x,y,z,zlim=c(0,20),xlim=c(5.8,8.5),ylim=c(-8.7,-6.0),
               theta=-10,phi=50,shade=1e-02,border=NULL,
               lwd=0.25,xlab="LnR0",ylab="Lnq",zlab="-veLL",
               scale=TRUE,expand=0.4,axes=TRUE,ticktype="detailed")
}
\keyword{datasets}
