% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspm.r
\name{plotASPM}
\alias{plotASPM}
\title{plotASPM plots catch, CPUE, Spawning Biomass and Harvest Rate}
\usage{
plotASPM(infish, CI = NA, defineplot = TRUE, target = 0.48, usef = 7, png = "")
}
\arguments{
\item{infish}{an object generated by the dynamics function}

\item{CI}{defaults to NA, if confidence intervals around the cpue have been
obtained using getLNCI, then the resulting matrix will generate 95pc CIs}

\item{defineplot}{define the plot size and character outside the plot or
automatically inside. Defaults to TRUE}

\item{target}{target depletion level. Defaults to 0.48}

\item{usef}{defines the font to use usef(ont),default = 7 bold times}

\item{png}{save a png file with the name in 'png', default = "", which
means no file produced}
}
\value{
Nothing, but it does plot six graphs in a single plot.
}
\description{
plotASPM after running fitASPM the optimum parameters can be
put through the dynamics function to generate a dataframe containing
the optimum dynamics. These can be plotted using plotASPM, which plots
out the catches, the Spawning Biomass, the relative CPUE and its fit to
the observed CPUE, and the harvest rate. This routine is still under
development to include more options.
}
\examples{
\dontrun{
data(westroughy)
fish <- westroughy$fish
glb <- westroughy$glb
props <- westroughy$props
pars <- c(7.1,-1,-7.7)  # logR0 logceSE logavq
bestaspm <- fitASPM(pars,minfun=dynamicsH,infish=fish,
                   inglb=glb,inprops=props)
out <- dynamicsH(bestaspm$estimate,fish,glb,props,full=TRUE)
plotASPM(out$fishery,defineplot=TRUE)
ceCI <- getLNCI(out$fishery[,"predCE"],exp(bestspm$estimate[2]))
plotASPM(out$fishery,CI=ceCI)
}  # infish=fisheryPen; CI=ceCI; defineplot=TRUE; target=0.48; usef=7;png=""
# infish=outH$fishery; CI=ceCI;defineplot=TRUE; target=0.48; usef=7;png=""
}
