% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.R
\name{plotdiffcomp}
\alias{plotdiffcomp}
\title{plotcompfit plots observed vs predicted composition proportions}
\usage{
plotdiffcomp(
  diffcomp,
  analysis = "deviations",
  ylabel = "opt49_ages",
  console = TRUE,
  outdir = "",
  cex = 1,
  topcex = 0.9,
  relmin = FALSE
)
}
\arguments{
\item{diffcomp}{the difference between the observed composition data and the
predicted. Dervied from the plotcompfit function
with ages or sizes as rownames and years as column names}

\item{analysis}{a character name for the comparison being made,
default='deviations'. Whatever is used should have no spaces as it is
also used in the filename, if used}

\item{ylabel}{a character label for the y-axis, this is added to the
contents of the analysis argument, default='opt49_ages'}

\item{console}{should the plot got to the console or file, default = TRUE}

\item{outdir}{the directory into which to send the file is console = FALSE,
default =''}

\item{cex}{the generic size of the font used in the plots}

\item{topcex}{the font size used for the top labels = year and sample size}

\item{relmin}{should the absolute difference, topright text, be relative to
the minimum relative difference or teh actual relative difference,
default = FALSE, giving the real absolute difference}
}
\value{
It generates a plot and, invisibly, a list of the filename and
caption in case it is saved as a file.
}
\description{
plotcompfit generates a plot of the predicted proportional
composition in the catches versus the observed proportional composition
data to illustrate the fit of the model to the observed data. The xlabel
of all plots is left blank, but the range of each x-axis is printed at
the bottom of all.
}
\examples{
\dontrun{
  diffcomp <- compfit$obscomp - compfit$predcomp  # from plotcompfit
  diffout <- plotdiffcomp(diffcomp,analysis="composition_deviations",
                          ylabel="opt49_age",console=TRUE,outdir="",
                          cex=1.0,topcex=0.9) 
}
#  diffcomp;analysis="composition_deviations";ylabel="opt49_age"
#  console=TRUE;outdir="";cex=1.0;topcex=0.9
}
