% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.R
\name{plotdynfish}
\alias{plotdynfish}
\title{plotdynfish generates a plot of a fisheries dynamics}
\usage{
plotdynfish(
  outfish,
  console = TRUE,
  addtitle = "",
  prepplot = TRUE,
  rundir = "",
  width = 8,
  height = 7,
  nfleet = 2,
  obsdata = TRUE,
  year = "year",
  recruit = c("recruit", "predrec"),
  depl = "deplsB",
  gears = c("Trawl", "Autoline"),
  catch = c("twl", "auln"),
  instF = c("twlPF", "aulnPF"),
  cecols = c("twlPCE", "twlCE", "aulnPCE", "aulnCE")
)
}
\arguments{
\item{outfish}{a data.frame of the fishery dynamics, as a minimum it must
obviously include observed and predicted catches and CPUE, spawning
biomass depletion, recruitment, and instantaneous F values. The column
headings of which can be identified in teh columns argument, in which
the order is important.}

\item{console}{should the plot go to the console, default = TRUE}

\item{addtitle}{The default filename if 'fishery_dynamics.png' use addtitle
to add text in front of that. eg addtitle='Speciesname_'}

\item{prepplot}{should the plotprep function be used? default = TRUE}

\item{rundir}{the directory into which to save the file if console=FALSE,
default = ''}

\item{width}{the width of the plot within plotprep if used, default=8}

\item{height}{the height of the plot within plotprep if used, default=7}

\item{nfleet}{default = 2, determines how many gears are expected can only
be 1 or 2. If nfleet = 1 consider reducing height}

\item{obsdata}{default = TRUE. To plot CPUE residuals the observed CPUE is
required. When plotting simulated data there are no 'observed' so set
obsdata to FALSE, which omits the residual plots.}

\item{year}{character name of the year column, default ='year'}

\item{recruit}{character name of the recruitment column, the predrec column
contains the predicted recruitment without the deviate value. default =
'recruit' and 'predrec'. If no predrec values leave as default}

\item{depl}{character name of the spawning biomass depletion column default
= 'deplsB'}

\item{gears}{character names of the fishing gear names, not columns.
default = 'Trawl' and 'Autoline'}

\item{catch}{character names of the catches columns. default = 'twl',"auln}

\item{instF}{character name of the ycolumns containing the estimates of
instantaneous F for each gear through time. default='twlPF', 'aulnPF"}

\item{cecols}{character name of the column of cpue information, predicted
values for each gear followed by the observed values for that gear (if
present). default = 'twlPCE','twlCE','aulnPCE','aulnCE'}
}
\value{
nothing but it does generate a plot
}
\description{
plotdynfish generates a plot of a fisheries dynamics. This
includes plots of predicted CPUE vs observed CPUE, the residuals for the
same (separate plots for different gears), the spawning biomass
depletion, the recruitment levels, the catches, and the instantaneous
F levels
}
\examples{
print("Wait on example data - again")
print("Can now plot simulated dynamics")
}
