% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agestruct.R
\name{plotdynfish}
\alias{plotdynfish}
\title{plotdynfish generates a plot of a fisheries dynamics}
\usage{
plotdynfish(
  outfish,
  console = TRUE,
  addtitle = "",
  prepplot = TRUE,
  rundir = "",
  width = 8,
  height = 7,
  nfleet = 2,
  columns = c("year", "twlC", "aulnC", "twlCE", "aulnCE", "twlPCE", "aulnPCE", "deplsB",
    "recruit", "twlPF", "aulnPF", "Trawl", "Autoline")
)
}
\arguments{
\item{outfish}{a data.frame of the fishery dynamics, as a minimum it must
obviously include observed and predicted catches and CPUE, spawning
biomass depletion, recruitment, and instantaneous F values. The column
headings of which can be identified in teh columns argument, in which
the order is important.}

\item{console}{should the plot go to the console, default = TRUE}

\item{addtitle}{The default filename if 'fishery_dynamics.png' use addtitle
to add text in front of that. eg addtitle='Speciesname_'}

\item{prepplot}{should the plotprep function be used? default = TRUE}

\item{rundir}{the directory into which to save the file if console=FALSE,
default = ''}

\item{width}{the width of the plot within plotprep if used, default=8}

\item{height}{the height of the plot within plotprep if used, default=7}

\item{nfleet}{default = 2, determines how many gears are expected can only
be 1 or 2. If nfleet = 1 consider reducing height}

\item{columns}{column headings identifying the components to be plotted. The
default is: columns=c('year','twlC','aulnC','twlCE','aulnCE', 'twlPCE',
'aulnPCE','deplsB','recruit','twlPF','aulnPF','Trawl','Autoline'), which
identify the year column, the catch columns, the observed CPUE columns,
the predicted CE cols, the depletion and recruitment cols, then the
instantaneous F estimates by gear, and finally the gear names. If nfleet
= 1 then, obviously one only lists the single gear.}
}
\value{
nothing but it does generate a plot
}
\description{
plotdynfish generates a plot of a fisheries dynamics. This
includes plots of predicted CPUE vs observed CPUE, the residuals for the
same (separate plots for different gears), the spawning biomass
depletion, the recruitment levels, the catches, and the instantaneous
F levels
}
\examples{
print("Wait on example data - again")
}
