% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspm.r
\name{plotprops}
\alias{plotprops}
\title{plotprops generates a 2x2 plot of the fishery properties}
\usage{
plotprops(rundir, props, console = TRUE)
}
\arguments{
\item{rundir}{the directory in which the analysis is being run}

\item{props}{the matrix of fishery properties including laa, waa, maa, and
sela in columns 2 - 5}

\item{console}{should the plot go to the console or be saved as a png file
into rundir? default=TRUE ie plot to console}
}
\value{
nothing but it does generate a plot
}
\description{
plotprops generates a 2 x 2 plot of the fishery properties of
the length-at-age, maturity-at-age, weight-at-age, and selectivity-at-age
}
\examples{
require(hplot)
data("westroughy")
plotprops(rundir="",westroughy$props,console=TRUE)
}
