% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{plottmbprof}
\alias{plottmbprof}
\title{plottmbprof a wrapper function for the tmbprofile plot}
\usage{
plottmbprof(inmod, parname, CI = 0.95, adjust = 0.05)
}
\arguments{
\item{inmod}{the RTMB model from MakeADFun}

\item{parname}{the character name of the parameter from the model, the
column name of the variable of parameter being profiled}

\item{CI}{default = 0.95 the probability level of the Confidence intervals}

\item{adjust}{by how much should the labels be adjusted down and up,
default = 0.05 = 5\% of the y-scale}
}
\value{
the CI
}
\description{
plottmbprof provides a simplified interface to the RTMB
tmbprofile plot. It plots the profile values and adds the optimum and
CI valued confidence intervals.
}
\examples{
print("wait on examples")
# syntax:  plottmbprof(model,"logR0",CI=0.95,adjust=0.05)
# inmod=model; parname="logR0"; CI=0.95
}
