% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agestruct.R
\name{simdynF}
\alias{simdynF}
\title{simdynF describe the ASPM dynamics}
\usage{
simdynF(glb, stk, reps = 6, full = FALSE)
}
\arguments{
\item{glb}{the glb data.frame from readdata or built in dataset const2.}

\item{stk}{the stock object from makestock}

\item{reps}{how many iterations to use in findF, default = 6}

\item{full}{should all outputs from dynamics be given. When fitting the
model, set this to FALSE. Once fitted, change this to TRUE to get all
the required outputs.}
}
\value{
if !full then a data.frame containing the fishery dynamics according
to the input arguments glb and stk. Includes Year, Catch, PredC, SpawnB,
ExploitB, FullH, CPUE, PredCE, Deplete, Recruit, FullF. if full, then a
list of the fishery dynamics, plus Nt, NumC, catchN, Lt, LC, and LCflt.
}
\description{
simdynF summarizes the dynamics of an Age-Structured
Integrated Model (ASM). Fishing mortality is implemented as instantaneous
rates rather than as annual harvest rates. Nt is the numbers-at-age at
the start (or end) of a year, catchN is the predicted numbers-at-age in
the catch of each fleet, NumC is the total numbers-at-age in the catch,
Lt is the predicted numbers-at-size at the start of each year, Gtran\emph{Nt,
LCflt is the numbers-at-size in the catch for each fleet, Gtran}catchN,
at the end of each year, and LC is the total numbers-at-size in the
catch, which is Gtran * NumC. These numbers-at-size ignore the fact that
larger fish are differentially taken by size due to selectivity (a weak
assumption in purely age-structured models.
}
\examples{
data(simconst2)
glb <- initiateglobals(simconst2,agesorsizes="ages",selabove=1)
stock <- makesimstock(glb=glb)
out <- simdynF(glb=glb,stk=stock,reps=6,full=TRUE)
str(out)
print(round(out$fishery,3)) 
}
