% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agestruct.R
\name{templateSIM}
\alias{templateSIM}
\title{templateSIM generates template for simulating stock dynamics}
\usage{
templateSIM(rundir, filename = "F1-A1-S1-age.csv", nfleet = 1)
}
\arguments{
\item{rundir}{directory in which to find the data file and run the analysis}

\item{filename}{the name of the data file to be produced, default=
'F1-A1-S1-age.csv'.}

\item{nfleet}{the number of fleets for which data will be generated,
default = 1.}
}
\value{
the function write a data file to rundir and returns the filename
}
\description{
templateSIM writes a data CSV file for either a 1- or 2-fleet
stock ready for simulating the stock's dynamics. That uses the functions
getconstants, initiateglobals, and makesimstock. Then, after adjusting
the data file to obtain the required dynamics, use makedataset to
generate a data set for use in model fitting.
}
\examples{
\dontrun{
  rundir <- tempdir()
  dirExists(rundir,verbose=TRUE)
  templateSIM(rundir,filename="F1-A1-S1-age.csv",nfleet=1)
  dir(rundir)
}
}
