% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspm.r
\name{unfished}
\alias{unfished}
\title{unfished generates the numbers at age for an unfished population}
\usage{
unfished(glob, props, inR0)
}
\arguments{
\item{glob}{the global constants object containing biology and structure}

\item{props}{the data.frame containing laa, maa, waa, maa, and sela}

\item{inR0}{the log of unfished recruitment from B0}
}
\value{
a list containing N0, R0, A0, and B0
}
\description{
unfished generates the numbers at age for an unfished
population, and determines the recruitment dynamics. It requires the
input of R0. The output includes the unfished numbers-at-age N0 plus
B0, A0, and R0.
}
\examples{
data("westroughy")
glb <- westroughy$glb  # contains a guess at log(R0)
props <- westroughy$props
unfish <- unfished(glob=glb,props=props,inR0=glb$R0)
print(unfish)
}
