% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agestruct.R
\name{writedatafile}
\alias{writedatafile}
\title{writedatafile writes out a simulated data file to rundir/filename}
\usage{
writedatafile(const, glb, histC, rundir, filename, datatitle)
}
\arguments{
\item{const}{the object made by getconstants aimed at simulating dynamics}

\item{glb}{the globals object obtained from initiategeglobals}

\item{histC}{an object inside the output from getconstants}

\item{rundir}{the directory path to the subdirectory in which all analysis
data and results are stored.}

\item{filename}{the name of the data file, MUST be a '.csv' file}

\item{datatitle}{line 1 of data file decribing its contents, default = ''}
}
\value{
nothing but it does write a data file to rundir/filename
}
\description{
writedatafile writes out a simulated data file to
rundir/filename that is suitable for use with dynF2, ie only catches and
cpue. The intent is to add the facility to include age- and size-
composition data later.
}
\examples{
print("see help for makedatafile, which uses writedatafile")
}
